### 2025-07-14 #251 @pllim
astropy: A Community Python Library for Astronomy
Submitting Author: @pllim
All current maintainers: @astropy
Package Name: astropy
One-Line Description of Package: A Community Python Library for Astronomy
Repository Link: https://github.com/astropy/astropy
Version submitted: 7.1.0
Archive: 10.5281/zenodo.17756022
Version accepted: 7.1.0

---

The Astropy Project is a community effort to develop a core package for astronomy using the [Python programming language](http://www.python.org/) and improve usability, interoperability, and collaboration between astronomy Python packages. The core astropy package contains functionality aimed at professional astronomers and astrophysicists, but may be useful to anyone developing astronomy software. The Astropy Project also includes "affiliated packages," Python packages that are not necessarily developed by the core development team, but share the goals of Astropy, and often build from the core package's code and infrastructure.

	- [x] Scientific software wrappers (wcslib?)

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

  - Who is the target audience and what are scientific applications of this package?  Target audience are people involved in astrophysics or astronomy who use Python. astropy is the core library that the Astropy Project builds upon.

  - Are there other Python packages that accomplish the same thing? If so, how does yours differ? No.

- [ ] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`. (I can add that later; out of time today.)
- [x] The package is deposited in a long-term repository with the DOI: https://zenodo.org/records/15473616

and editors to fill this out. (I can fill this out later; out of time today.)
