### 2025-07-14 #251 @pllim
Note to self: sunpy did it here https://github.com/pyOpenSci/software-submission/issues/147

Side discussion: https://github.com/pyOpenSci/software-submission/pull/252
