### 2025-07-14 efe6d5f
Apply formatting changes suggested by ruff (with a few reservations!).
