### 2025-07-14 f8e0f5a
Apply formatting changes suggested by ruff (with a few reservations!).
