+++ astropy/nddata/mixins/tests/test_ndarithmetic.py
@@ -1394,7 +1394,6 @@ def test_nddata_bitmask_arithmetic():
     ],
 )
 def test_arithmetics_dtypes_with_scalar(data1, data2):
-
     # Addition
     out1 = data1.add(data2)
     ref1 = data1.data + data2
@@ -1423,30 +1422,17 @@ def test_arithmetics_dtypes_with_scalar(data1, data2):
     ("data1", "data2"),
     [
         (
-            NDDataRef(
-                np.array([1, 2, 3, 4], dtype=np.uint16),
-                unit=u.adu
-            ),
-            u.Quantity(2, dtype=np.uint16, unit=u.adu)
+            NDDataRef(np.array([1, 2, 3, 4], dtype=np.uint16), unit=u.adu),
+            u.Quantity(2, dtype=np.uint16, unit=u.adu),
         ),
         (
-            NDDataRef(
-                np.array([1., 2., 3., 4.], dtype=np.float32),
-                unit=u.adu
-            ),
-            u.Quantity(2.0, dtype=np.float32, unit=u.adu)
+            NDDataRef(np.array([1.0, 2.0, 3.0, 4.0], dtype=np.float32), unit=u.adu),
+            u.Quantity(2.0, dtype=np.float32, unit=u.adu),
         ),
-        (
-            NDDataRef(
-                np.array([1., 2., 3., 4.]),
-                unit=u.adu
-            ),
-            2.0 * u.adu
-        ),
+        (NDDataRef(np.array([1.0, 2.0, 3.0, 4.0]), unit=u.adu), 2.0 * u.adu),
     ],
 )
 def test_add_quantity_matching_dtype(data1, data2):
-
     # Addition
     out1 = data1.add(data2)
     assert_array_equal(out1.data, data1.data + data2.value)
@@ -1464,8 +1450,8 @@ def test_add_quantity_matching_dtype(data1, data2):
     "data1",
     [
         NDDataRef(np.array([1, 2, 3, 4], dtype=np.uint16), unit=u.adu),
-        NDDataRef(np.array([1., 2., 3., 4.], dtype=np.float32), unit=u.adu),
-        NDDataRef(np.array([1., 2., 3., 4.]), unit=u.adu),
+        NDDataRef(np.array([1.0, 2.0, 3.0, 4.0], dtype=np.float32), unit=u.adu),
+        NDDataRef(np.array([1.0, 2.0, 3.0, 4.0]), unit=u.adu),
     ],
 )
 @pytest.mark.parametrize(
@@ -1479,7 +1465,6 @@ def test_add_quantity_matching_dtype(data1, data2):
     ],
 )
 def test_scale_dtypes_with_units(data1, data2):
-
     # Multiplication
     out1 = data1.multiply(data2)
     ref1 = data1.data * data2
@@ -1503,30 +1488,28 @@ def generate_simple_ndds_with_uncert_mask(nout=1):
                 uncertainty=VarianceUncertainty(
                     np.array([1, 2, 3, 4], dtype=np.uint16)
                 ),
-                mask=np.array([0, 1, 0, 0], dtype=np.uint8)
+                mask=np.array([0, 1, 0, 0], dtype=np.uint8),
             ),
             NDDataRef(
-                np.array([1., 2., 3., 4.], dtype=np.float32),
+                np.array([1.0, 2.0, 3.0, 4.0], dtype=np.float32),
                 uncertainty=StdDevUncertainty(
-                    np.array([1., 1.41, 1.73, 2.], dtype=np.float32)
+                    np.array([1.0, 1.41, 1.73, 2.0], dtype=np.float32)
                 ),
-                mask=np.array([0, 0, 1, 0], dtype=np.uint16)
[…]
