### 2025-07-15 c748950
Generic Table conversion through narhwals
