+++ pyproject.toml
@@ -87,6 +87,8 @@ all = [
     "html5lib>=1.1",
     "bleach>=3.2.1",
     "pandas>=2.0",
+    "polars>=1.31",
+    "narwhals>=1.46",
     "sortedcontainers>=1.5.7", # (older versions may work)
     "pytz>=2016.10", # (older versions may work)
     "jplephem>=2.15",
