### 2025-07-15 e7218ab
Generic Table conversion through narhwals
