+++ astropy/table/table.py
@@ -89,7 +89,9 @@ __doctest_skip__ = [
     "Table.convert_unicode_to_bytestring",
 ]
 
-__doctest_requires__ = {("Table.from_pandas", "Table.to_pandas"): ["pandas"]}
+__doctest_requires__ = {
+    ("Table.from_df", "Table.to_df"): ["pandas", "polars"],
+}
 
 _pprint_docs = """
     {__doc__}
@@ -4001,209 +4003,228 @@ class Table:
         """
         return groups.table_group_by(self, keys)
 
+    def _encode_mixins(tbl):
+        """Encode mixin columns to basic columns for Pandas/Polars compatibility."""
+        from astropy.time import TimeBase, TimeDelta
+
+        from . import serialize
+
+        time_cols = [col for col in tbl.itercols() if isinstance(col, TimeBase)]
+        if time_cols:
+            new_cols = []
+            for col in tbl.itercols():
+                new_col = col_copy(col, copy_indices=False) if col.info.indices else col
+                new_cols.append(new_col)
+            tbl = tbl.__class__(new_cols, copy=False)
+
+            for col in tbl.itercols():
+                col.info.indices.clear()
+
+            for col in time_cols:
+                if isinstance(col, TimeDelta):
+                    new_col = (col.sec * 1e9).astype("timedelta64[ns]")
+                    nat = np.timedelta64("NaT")
+                else:
+                    new_col = col.datetime64.copy()
+                    nat = np.datetime64("NaT")
+                if col.masked:
+                    new_col[col.mask] = nat
+                tbl[col.info.name] = new_col
+
+        # Convert the table to one with no mixins, only Column objects.
+        encode_tbl = serialize.represent_mixins_as_columns(tbl)
+        return encode_tbl
+
+    @deprecated(since="7.3", alternative="t.to_df('pandas')", name="t.to_pandas()")
     def to_pandas(self, index=None, use_nullable_int=True):
+        return self.to_df(
+            backend="pandas", index=index, use_nullable_int=use_nullable_int
+        )
+
+    def to_df(self, backend="pandas", index=None, use_nullable_int=True):
         """
-        Return a :class:`pandas.DataFrame` instance.
+        Convert the table to an eager DataFrame using the ``narwhals`` backend.
 
-        The index of the created DataFrame is controlled by the ``index``
-        argument.  For ``index=True`` or the default ``None``, an index will be
-        specified for the DataFrame if there is a primary key index on the
-        Table *and* if it corresponds to a single column.  If ``index=False``
-        then no DataFrame index will be specified.  If ``index`` is the name of
-        a column in the table then that will be the DataFrame index.
+        This method supports converting Astropy Table (or QTable) instances into a variety of
+        DataFrame formats via the ``narwhals`` library. The output can be any supported eager
+        DataFrame type, such as `pandas`, ``polars``, or others, depending on the specified
+        backend.
 
-        In addition to vanilla columns or masked columns, this supports Table
-        mixin columns like Quantity, Time, or SkyCoord.  In many cases these
-        objects have no analog in pandas and will be converted to a "encoded"
-        representation using only Column or MaskedColumn.  The exception is
-        Time or TimeDelta columns, which will be converted to the corresponding
-        representation in pandas using ``np.datetime64`` or ``np.timedelta64``.
-        See the example below.
+        Mixin columns such as ``Quantity``, ``Time``, ``TimeDelta``, or ``SkyCoord`` are automatically
+        converted to plain Column or MaskedColumn types when necessary. Time-related mixins
+        will be represented using ``np.datetime64`` or ``np.timedelta64`` as appropriate. All
+        other mixins are serialized into a flat column structure using standard representations.
 
         Parameters
         ----------
+        backend : str or module or ``narwhals.Implementation``, default="pandas"
[…]
