### 2025-07-15 35f71b2
Generic Table conversion through narhwals
