+++ astropy/table/tests/test_table.py
@@ -30,7 +30,7 @@ from astropy.table import (
 from astropy.tests.helper import assert_follows_unicode_guidelines
 from astropy.time import Time, TimeDelta
 from astropy.utils.compat import NUMPY_LT_1_25
-from astropy.utils.compat.optional_deps import HAS_PANDAS
+from astropy.utils.compat.optional_deps import HAS_NARWHALS, HAS_PANDAS, HAS_POLARS
 from astropy.utils.data import get_pkg_data_filename
 from astropy.utils.exceptions import AstropyDeprecationWarning, AstropyUserWarning
 from astropy.utils.metadata.tests.test_metadata import MetaBaseTest
@@ -2064,6 +2064,7 @@ def test_table_init_from_degenerate_arrays(table_types):
 
 
 @pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")
+@pytest.mark.skipif(not HAS_NARWHALS, reason="requires narwhals")
 class TestPandas:
     def test_simple(self):
         t = table.Table()
@@ -2104,7 +2105,7 @@ class TestPandas:
         d[["<i4", ">i4"]]
         d[["<f4", ">f4"]]
 
-        t2 = table.Table.from_pandas(d)
+        t2 = table.Table.from_df(d)
 
         for column in t.columns:
             if column in ("u", "s"):
@@ -2125,7 +2126,7 @@ class TestPandas:
         df = t.to_pandas()
         pd_dtype = np_dtype.replace("i", "I").replace("u", "U")
         assert str(df["col0"].dtype) == pd_dtype
-        t2 = Table.from_pandas(df)
+        t2 = Table.from_df(df)
         assert str(t2["col0"].dtype) == np_dtype
         assert np.all(t2["col0"].mask == [False, True])
         assert np.all(t2["col0"] == c)
@@ -2136,9 +2137,9 @@ class TestPandas:
         t["b"] = np.ones((3, 2))
 
         with pytest.raises(
-            ValueError, match="Cannot convert a table with multidimensional columns"
+            ValueError, match="Per-column arrays must each be 1-dimensional"
         ):
-            t.to_pandas()
+            t.to_df("pandas")
 
     def test_mixin_pandas(self):
         t = table.QTable()
@@ -2149,7 +2150,7 @@ class TestPandas:
         t["dt"] = TimeDelta([0, 2, 4, 6], format="sec")
 
         tp = t.to_pandas()
-        t2 = table.Table.from_pandas(tp)
+        t2 = table.Table.from_df(tp)
 
         assert np.allclose(t2["quantity"], [0, 1, 2, 3])
         assert np.allclose(t2["longitude"], [0.0, 1.0, 5.0, 6.0])
@@ -2227,7 +2228,7 @@ class TestPandas:
 
         with pytest.raises(ValueError) as err:
             t.to_pandas(index="not a column")
-        assert "index must be None, False" in str(err.value)
+        assert "is not in the table columns" in str(err.value)
 
     def test_mixin_pandas_masked(self):
         tm = Time([1, 2, 3], format="cxcsec")
@@ -2240,7 +2241,7 @@ class TestPandas:
         assert np.all(tp["tm"].isnull() == [False, True, False])
         assert np.all(tp["dt"].isnull() == [False, True, False])
 
-        t2 = table.Table.from_pandas(tp)
+        t2 = table.Table.from_df(tp)
 
         assert np.all(t2["tm"].mask == tm.mask)
         assert np.ma.allclose(t2["tm"].jd, tm.jd, rtol=1e-14, atol=1e-14)
@@ -2248,16 +2249,16 @@ class TestPandas:
[…]
