### 2025-07-15 a46972e
Generic Table conversion through narhwals
