### 2025-07-15 f1f592d
Generic Table conversion through narhwals
