+++ astropy/utils/compat/optional_deps.py
@@ -30,7 +30,9 @@ _optional_deps = [
     "lxml",
     "matplotlib",
     "mpmath",
+    "narwhals",
     "pandas",
+    "polars",
     "PIL",
     "pytz",
     "s3fs",
