+++ docs/changes/table/18435.feature.rst
@@ -1,0 +1,2 @@
+Added generic ``from_df`` and ``to_df`` methods to ``astropy.Table``.
+Deprecated ``from_pandas`` and ``to_pandas``.
