### 2025-07-15 85d7483
Improvements to C code, and expand tests to use a wider range of input values
