### 2025-07-15 13cbdf3
Add missing return
