+++ astropy/table/table.py
@@ -4059,13 +4059,144 @@ class Table:
         encode_tbl = serialize.represent_mixins_as_columns(tbl)
         return encode_tbl
 
-    @deprecated(since="7.3", alternative="t.to_df('pandas')", name="t.to_pandas()")
     def to_pandas(self, index=None, use_nullable_int=True):
-        return self.to_df(
-            backend="pandas", index=index, use_nullable_int=use_nullable_int
-        )
+        """
+        Return a :class:`pandas.DataFrame` instance.
 
-    def to_df(self, backend="pandas", index=None, use_nullable_int=True):
+        The index of the created DataFrame is controlled by the ``index``
+        argument.  For ``index=True`` or the default ``None``, an index will be
+        specified for the DataFrame if there is a primary key index on the
+        Table *and* if it corresponds to a single column.  If ``index=False``
+        then no DataFrame index will be specified.  If ``index`` is the name of
+        a column in the table then that will be the DataFrame index.
+
+        In addition to vanilla columns or masked columns, this supports Table
+        mixin columns like Quantity, Time, or SkyCoord.  In many cases these
+        objects have no analog in pandas and will be converted to a "encoded"
+        representation using only Column or MaskedColumn.  The exception is
+        Time or TimeDelta columns, which will be converted to the corresponding
+        representation in pandas using ``np.datetime64`` or ``np.timedelta64``.
+        See the example below.
+
+        Parameters
+        ----------
+        index : None, bool, str
+            Specify DataFrame index mode
+        use_nullable_int : bool, default=True
+            Convert integer MaskedColumn to pandas nullable integer type.  If
+            ``use_nullable_int=False`` then the column is converted to float
+            with NaN.
+
+        Returns
+        -------
+        dataframe : :class:`pandas.DataFrame`
+            A pandas :class:`pandas.DataFrame` instance
+
+        Raises
+        ------
+        ImportError
+            If pandas is not installed
+        ValueError
+            If the Table has multi-dimensional columns
+
+        Examples
+        --------
+        Here we convert a table with a few mixins to a
+        :class:`pandas.DataFrame` instance.
+
+          >>> import pandas as pd
+          >>> from astropy.table import QTable
+          >>> import astropy.units as u
+          >>> from astropy.time import Time, TimeDelta
+          >>> from astropy.coordinates import SkyCoord
+
+          >>> q = [1, 2] * u.m
+          >>> tm = Time([1998, 2002], format='jyear')
+          >>> sc = SkyCoord([5, 6], [7, 8], unit='deg')
+          >>> dt = TimeDelta([3, 200] * u.s)
+
+          >>> t = QTable([q, tm, sc, dt], names=['q', 'tm', 'sc', 'dt'])
+
+          >>> df = t.to_pandas(index='tm')
+          >>> with pd.option_context('display.max_columns', 20):
+          ...     print(df)
+                        q  sc.ra  sc.dec              dt
+          tm
+          1998-01-01  1.0    5.0     7.0 0 days 00:00:03
+          2002-01-01  2.0    6.0     8.0 0 days 00:03:20
+
+        """
+        from pandas import DataFrame, Series
+
+        if index is not False:
+            if index in (None, True):
+                # Default is to use the table primary key if available and a single column
+                if self.primary_key and len(self.primary_key) == 1:
+                    index = self.primary_key[0]
+                else:
+                    index = False
+            else:
+                if index not in self.colnames:
+                    raise ValueError(
+                        "index must be None, False, True or a table column name"
+                    )
+
+        tbl = self._encode_mixins()
+
+        badcols = [name for name, col in self.columns.items() if len(col.shape) > 1]
+        if badcols:
+            raise ValueError(
[…]
