### 2025-07-15 e31b372
Fixed documentation spelling/whitespace
