### 2025-07-15 e00b50c
Output CUNIT in WCS __repr__, and add documentation and what's new entry
