+++ docs/whatsnew/7.2.rst
@@ -72,3 +72,50 @@ Here is an example of how to use the
 
 By combining these traits, you can easily construct custom cosmology classes with
 precisely the features you need, without having to reimplement common functionality.
+
+Preserving units in FITS-WCS
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+By default, the :class:`~astropy.wcs.WCS` class always converts units into degrees
+for angles, and SI units for other physical types::
+
+    >>> from astropy.io import fits
+    >>> from astropy.wcs import WCS
+    >>> header = """
+    ... CTYPE1  = 'GLON-CAR'
+    ... CTYPE2  = 'GLAT-CAR'
+    ... CTYPE3  = 'FREQ'
+    ... CUNIT1  = 'arcsec'
+    ... CUNIT2  = 'arcsec'
+    ... CUNIT3  = 'GHz'
+    ... CRVAL1  = 10
+    ... CRVAL2  = 20
+    ... CRVAL3  = 50
+    """.strip()
+    >>> wcs = WCS(fits.Header.fromstring(header, sep='\n'))
+    >>> wcs  # doctest: +FLOAT_CMP
+    WCS Keywords
+
+    Number of WCS axes: 3
+    CTYPE : 'GLON-CAR' 'GLAT-CAR' 'FREQ'
+    CUNIT : deg deg Hz
+    CRVAL : 0.002777777777777778 0.005555555555555556 50000000000.0
+    ...
+
+However, it is now possible to preserve the original units by specifying
+``preserve_units=True`` when initializing the :class:`~astropy.wcs.WCS`
+object::
+
+    >>> wcs = WCS(fits.Header.fromstring(header, sep='\n'), preserve_units=True)
+    >>> wcs  # doctest: +FLOAT_CMP
+    WCS Keywords
+
+    Number of WCS axes: 3
+    CTYPE : 'GLON-CAR' 'GLAT-CAR' 'FREQ'
+    CUNIT : arcsec arcsec GHz
+    CRVAL : 10.0 20.0 50.0
+    ...
+
+When using this, any input/output world coordinates will now be in these
+units, and accessing any of the parameters such as ``wcs.wcs.crval`` will
+return values in the original header units.
