### 2025-07-15 7658447
Output CUNIT in WCS __repr__, and add documentation and what's new entry
