### 2025-07-15 dd730c5
Unified index parameter in to_pandas and to_df
