### 2025-07-15 20915fd
Unified index parameter in to_pandas and to_df
