### 2025-07-15 1255685
Fix doctests and remove trailing spaces in WCS.__repr__
