+++ docs/wcs/wcsapi.rst
@@ -253,7 +253,7 @@ for angles, and SI units for other physical types::
     >>> wcs = WCS(fits.Header.fromstring(header, sep='\n'))
     >>> wcs  # doctest: +FLOAT_CMP
     WCS Keywords
-
+    <BLANKLINE>
     Number of WCS axes: 3
     CTYPE : 'GLON-CAR' 'GLAT-CAR' 'FREQ'
     CUNIT : deg deg Hz
@@ -263,7 +263,7 @@ for angles, and SI units for other physical types::
     PC2_1 PC2_2 PC2_3  : 0.0 1.0 0.0
     PC3_1 PC3_2 PC3_3  : 0.0 0.0 1.0
     CDELT : 0.0002777777777777778 0.0002777777777777778 1000000000.0
-    NAXIS : 0  0
+    NAXIS : 0  0  0
 
 However, it is possible to preserve the original units by specifying
 ``preserve_units=True`` when initializing the :class:`~astropy.wcs.WCS`
@@ -272,7 +272,7 @@ object::
     >>> wcs = WCS(fits.Header.fromstring(header, sep='\n'), preserve_units=True)
     >>> wcs  # doctest: +FLOAT_CMP
     WCS Keywords
-
+    <BLANKLINE>
     Number of WCS axes: 3
     CTYPE : 'GLON-CAR' 'GLAT-CAR' 'FREQ'
     CUNIT : arcsec arcsec GHz
@@ -282,7 +282,7 @@ object::
     PC2_1 PC2_2 PC2_3  : 0.0 1.0 0.0
     PC3_1 PC3_2 PC3_3  : 0.0 0.0 1.0
     CDELT : 1.0 1.0 1.0
-    NAXIS : 0  0
+    NAXIS : 0  0  0
 
 When using this, any input/output world coordinates will now be in these
 units, and accessing any of the parameters such as ``wcs.wcs.crval`` will
