### 2025-07-15 b894184
Fix doctests and remove trailing spaces in WCS.__repr__
