### 2025-07-15 416977d
Applied ruff formatting
