### 2025-07-15 4177679
Applied ruff formatting
