+++ astropy/wcs/tests/test_wcs.py
@@ -2256,16 +2256,16 @@ RADESYS = 'ICRS'               / Equatorial coordinate system
             "WCS Keywords\n"
             "\n"
             "Number of WCS axes: 5\n"
-            "CTYPE : 'RA---TAN' 'FREQ' 'DEC--TAN' 'WAVE' 'OFFSET' \n"
-            "CUNIT : arcsec GHz arcsec nm arcmin \n"
-            "CRVAL : 4.0 5.0 6.0 7.0 8.0 \n"
-            "CRPIX : 1.0 2.0 3.0 4.0 5.0 \n"
-            "PC1_1 PC1_2 PC1_3 PC1_4 PC1_5  : 1.0 0.0 0.0 0.0 0.0 \n"
-            "PC2_1 PC2_2 PC2_3 PC2_4 PC2_5  : 0.0 1.0 0.0 0.0 0.0 \n"
-            "PC3_1 PC3_2 PC3_3 PC3_4 PC3_5  : 0.0 0.0 1.0 0.0 0.0 \n"
-            "PC4_1 PC4_2 PC4_3 PC4_4 PC4_5  : 0.0 0.0 0.0 1.0 0.0 \n"
-            "PC5_1 PC5_2 PC5_3 PC5_4 PC5_5  : 0.0 0.0 0.0 0.0 1.0 \n"
-            "CDELT : 4.0 3.0 2.0 1.0 6.0 \n"
+            "CTYPE : 'RA---TAN' 'FREQ' 'DEC--TAN' 'WAVE' 'OFFSET'\n"
+            "CUNIT : arcsec GHz arcsec nm arcmin\n"
+            "CRVAL : 4.0 5.0 6.0 7.0 8.0\n"
+            "CRPIX : 1.0 2.0 3.0 4.0 5.0\n"
+            "PC1_1 PC1_2 PC1_3 PC1_4 PC1_5  : 1.0 0.0 0.0 0.0 0.0\n"
+            "PC2_1 PC2_2 PC2_3 PC2_4 PC2_5  : 0.0 1.0 0.0 0.0 0.0\n"
+            "PC3_1 PC3_2 PC3_3 PC3_4 PC3_5  : 0.0 0.0 1.0 0.0 0.0\n"
+            "PC4_1 PC4_2 PC4_3 PC4_4 PC4_5  : 0.0 0.0 0.0 1.0 0.0\n"
+            "PC5_1 PC5_2 PC5_3 PC5_4 PC5_5  : 0.0 0.0 0.0 0.0 1.0\n"
+            "CDELT : 4.0 3.0 2.0 1.0 6.0\n"
             "NAXIS : 0  0  0  0  0"
         )
         assert str(self.wcs_preserve) == expected
