### 2025-07-15 151dcbf
Changed use_nullable_int=False behaviour to raise an error with masked int columns
