### 2025-07-15 1e2e157
Changed use_nullable_int=False behaviour to raise an error with masked int columns
