### 2025-07-15 cbe7160
Fixed WCS.__repr__
