### 2025-07-15 04bf0de
Backport PR #18434: Ensure Python files without a shebang are not executable
