### 2025-07-15 a428454
Avoid modifying input world coordinates in-place
