### 2025-07-15 3062d31
Added regression test for in-place modifications to world array, and started adding -TAB test
