### 2025-07-15 813e116
Check ruff B007 in time and fix issues
