### 2025-07-15 22d862f
Check ruff B007 in time and fix issues
