+++ astropy/visualization/wcsaxes/tests/test_coordinate_helpers.py
@@ -328,6 +328,8 @@ def test_set_visible_invalid_type(invalid_input):
 
     with pytest.raises(TypeError):
         coord.set_visible(invalid_input)
+
+
 def test_set_position_invalid():
     fig = Figure()
     _canvas = FigureCanvasAgg(fig)
