### 2025-07-15 1561b29
Revert "Run pre-commit"
This reverts commit 45f577d6ba8ffcaef7bb1a42c04a9d57767df91c.
