### 2025-07-15 8ce6446
Revert "Merge remote-tracking branch 'upstream/main' into fix-18332"
This reverts commit d9f5fb40e8ab79bf5ac0dc7b38c0244df813e5ce, reversing
changes made to 9d51e46b89828f428038a4736c613c37cb929459.
