+++ astropy/visualization/wcsaxes/_auto.py
@@ -114,10 +114,9 @@ def auto_assign_coord_positions(ax):
 
     # Finalize assignments
     for coord, spine in zip(auto_coords, best_option):
-        position = [spine, "#"] if spine != " " else "#"
         if "#" in coord.get_ticks_position():
-            coord.set_ticks_position(position)
+            coord.set_ticks_position([spine, "#"])
         if "#" in coord.get_ticklabel_position():
-            coord.set_ticklabel_position(position)
+            coord.set_ticklabel_position([spine, "#"])
         if "#" in coord.get_axislabel_position():
-            coord.set_axislabel_position(position)
+            coord.set_axislabel_position([spine, "#"])
