+++ astropy/visualization/wcsaxes/ticks.py
@@ -105,7 +105,7 @@ class Ticks(Line2D):
             return 0.0
 
     def set_visible_axes(self, visible_axes):
-        self._visible_axes = self._frame._validate_positions(visible_axes)
+        self._visible_axes = visible_axes
 
     def get_visible_axes(self):
         if self._visible_axes == "all":
