### 2025-07-15 d1e46b8
Reapply "Merge remote-tracking branch 'upstream/main' into fix-18332"
This reverts commit 8ce64466c489832bfa7fbb7c1826bcf37c529af1.
