### 2025-07-15 27d1eb6
Try adding a changelog entry. Tweak ruff formatting slightly.
