+++ docs/changes/nddata/18392.bugfix.rst
@@ -1,0 +1,7 @@
+Fixed unexpected upcasting to 64 bits when doing arithmetic with Python scalars
+on ``numpy`` 2 (#18384).
+
+Don't upcast ``NDData``  to ``float64`` unnecessarily when doing arithmetic
+involving a single unit (consistent with the behaviour when there are no units).
+Upcasting still occurs if an operand's unit gets converted to match the other,
+or where required by the other operand's dtype.
