### 2025-07-15 dec6a07
Revert "Emit warning if axes given to set_*_position are invalid"
