### 2025-07-15 8570013
Removed warning for ICRS on SpectralCoord
