### 2025-07-15 a87b61a
Refactor visibility setter in set_visible()
