### 2025-07-16 56745ab
Add a private property to check unit scaling, and only do scaling if any of the scales are not 1. Also explicitly check the units of the TAB WCS
