+++ astropy/wcs/src/wcslib_wrap.c
@@ -1078,16 +1078,6 @@ int initialize_preserve_units(PyWcsprm* self) {
       }
     }
 
-    if (self->unit_scaling == NULL) {
-      self->unit_scaling = malloc(self->x.naxis * sizeof(double));
-      if (self->unit_scaling == NULL) {
-          PyErr_NoMemory();
-          return -1;
-      }
-      for (int i = 0; i < self->x.naxis; ++i) {
-          self->unit_scaling[i] = 1.0;
-      }
-    }
   }
 
   return 0;
@@ -1101,6 +1091,15 @@ int check_unit_changes(PyWcsprm* self) {
 
   // Check which units have changed, and store scaling if changed.
   if (self->original_cunit != NULL) {
+
+    if (self->unit_scaling == NULL) {
+      self->unit_scaling = malloc(self->x.naxis * sizeof(double));
+      if (self->unit_scaling == NULL) {
+          PyErr_NoMemory();
+          return -1;
+      }
+    }
+
     for (int i = 0; i < self->x.naxis; ++i) {
       if (strcmp(self->original_cunit[i], self->x.cunit[i]) != 0) {
         status = wcsunits(self->original_cunit[i], self->x.cunit[i], &scale, &offset, &power);
@@ -1115,7 +1114,28 @@ int check_unit_changes(PyWcsprm* self) {
                 );
           self->unit_scaling[i] = scale;
         }
+      } else {
+        self->unit_scaling[i] = 1.0;
       }
+
+    }
+
+    // We now check if all the scales are 1. If so, then we actually deallocate the array
+    // again to not waste time doing unit conversions subsequently. It might seem wasteful
+    // to allocate the array and then deallocate it, but the other option would be to
+    // allocate a temporary array for the scales in this function and then only if all the
+    // values are not 1 allocate self->unit_scaling, but then we'd be doing two allocations.
+
+    int scaling_needed = 0;
+    for (int i = 0; i < self->x.naxis; ++i) {
+        if (self->unit_scaling[i] != 1.0) {
+            scaling_needed = 1;
+            break;
+        }
+    }
+    if (!scaling_needed) {
+      free(self->unit_scaling);
+      self->unit_scaling = NULL;
     }
 
   }
@@ -1250,7 +1270,7 @@ PyWcsprm_get_cdelt_func(
 
   naxis = self->x.naxis;
 
-  if (self->original_cunit != NULL) {
+  if (self->unit_scaling != NULL) {
     result = get_scaled_double_array_readonly("cdelt", self->x.cdelt, self->unit_scaling, 1, &naxis, (PyObject*)self);
   } else {
     result = get_double_array_readonly("cdelt", self->x.cdelt, 1, &naxis, (PyObject*)self);
@@ -1627,7 +1647,7 @@ PyWcsprm_p2s(
   if (status == 0 || status == 8) {
     // Since the conversion succeeded, if user has requested to preserve units,
     // we convert the world coordinates to the original units
-    if (self->original_cunit != NULL) {
+    if (self->unit_scaling != NULL) {
       double *world_data = (double *)PyArray_DATA(world);
       for (npy_intp i = 0; i < nelem; ++i) {
         for (npy_intp j = 0; j < ncoord; ++j) {
@@ -1731,7 +1751,7 @@ PyWcsprm_s2p(
 
[…]
