### 2025-07-16 a5d8ab9
Add a private property to check unit scaling, and only do scaling if any of the scales are not 1. Also explicitly check the units of the TAB WCS
