### 2025-07-16 65207fc
Create new methods in coordinate_helpers.py
