### 2025-07-16 4d3f389
Turn the warning into a log debug
