### 2025-07-16 afcbc82
Update docs/changes/wcs/18440.api.rst
Co-authored-by: P. L. Lim <2090236+pllim@users.noreply.github.com>
