+++ docs/changes/io.fits/18379.bugfix.rst
@@ -1,1 +1,0 @@
-Allow passing fsspec parameters to the FITS Table reader.
