### 2025-07-17 429a28a
MNT: Minor refactoring of _get_differences
