### 2025-07-17 4cd59a6
Fix error handling for setting cunit
