### 2025-07-18 599ca54
Introduce `coordinates.is_rotation_or_reflection()`
The new function is essentially `matrix_utilities.is_O3()` renamed. The
old name was obscure to anyone not familiar enough with group theory to
recognize the O(3) group. The function is not documented under its old
name, but it is nonetheless mentioned in the change log, so the old name
is deprecated as a courtesy instead of being removed immediately.
