+++ docs/changes/coordinates/18418.api.rst
@@ -1,4 +1,15 @@
+The ``matrix_utilities`` module was not included in the ``astropy`` API
+documentation, but it was nonetheless explicitly referred to in some of the
+other documentation.
+This made it unclear if the functions in the module are public or private.
+The public matrix utilities ``is_rotation_or_reflection()`` and
+``rotation_matrix()`` have been made available from the ``astropy.coordinates``
+namespace and should be imported from there.
+Functions not available from the ``astropy.coordinate`` namespace are private
+and may be changed or removed without warning.
+However, three functions have been explicitly deprecated, despite being
+private, as a courtesy to existing users.
 ``matrix_utilites.angle_axis()`` and ``matrix_utilites.is_rotation()`` are
-deprecated.
-``matrix_utilities.is_O3()`` is renamed to ``is_rotation_or_reflection()`` and
-the old name is deprecated.
+deprecated without replacement.
+``matrix_utilities.is_O3()`` is deprecated and the public
+``is_rotation_or_reflection()`` function can be used as a replacement.
