### 2025-07-18 f13fcf6
Make matrix utilities available from `coordinates`
`is_rotation_or_reflection()` and `rotation_matrix()` are now available
from the top-level `coordinates` namespace.

`matrix_transpose()` is not made available from `coordinates` because
users more concerned about readability can use `np.matrix_transpose()`
(starting from `numpy` 2.0) and users more concerned about performance
should use the `swapaxes()` method directly.

The other matrix utilities are deprecated.
