+++ astropy/coordinates/tests/test_matrix_utilities.py
@@ -4,12 +4,11 @@ import pytest
 from numpy.testing import assert_allclose, assert_array_equal
 
 from astropy import units as u
+from astropy.coordinates import is_rotation_or_reflection, rotation_matrix
 from astropy.coordinates.matrix_utilities import (
     angle_axis,
     is_O3,
     is_rotation,
-    is_rotation_or_reflection,
-    rotation_matrix,
 )
 from astropy.utils.exceptions import AstropyDeprecationWarning
 
