### 2025-07-18 cd8c967
Initial commit of darkenergycomponent trait class
