### 2025-07-18 b0c012a
Addressing the darkenergy paramter review comment
