### 2025-07-18 1711c13
Update astropy/cosmology/_src/traits/darkenergy.py
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
