+++ docs/cosmology/traits.rst
@@ -10,9 +10,10 @@ The :mod:`~astropy.cosmology.traits` module provides reusable components, called
 :term:`traits <trait type>`, that encapsulate specific cosmological properties or
 behaviors. For example, the :class:`~astropy.cosmology.traits.HubbleParameter` trait
 provides the Hubble constant (``H0``) and related methods, while
-:class:`~astropy.cosmology.traits.ScaleFactor` and
-:class:`~astropy.cosmology.traits.TemperatureCMB` provide the scale factor and the
-temperature or the CMB, respectively.
+:class:`~astropy.cosmology.traits.ScaleFactor`,
+:class:`~astropy.cosmology.traits.TemperatureCMB` and
+:class:`~astropy.cosmology.traits.DarkEnergyComponent` provide the scale factor, the
+temperature or the CMB, and the Dark Energy component, respectively.
 
 By combining these traits, you can easily construct custom cosmology classes with
 precisely the features you need, without having to reimplement common functionality.
@@ -20,14 +21,15 @@ precisely the features you need, without having to reimplement common functional
 
 Here is an example of how to use the
 :class:`~astropy.cosmology.traits.HubbleParameter`,
-:class:`~astropy.cosmology.traits.ScaleFactor`, and
-:class:`~astropy.cosmology.traits.TemperatureCMB` traits in a custom cosmology class:
+:class:`~astropy.cosmology.traits.ScaleFactor`,
+:class:`~astropy.cosmology.traits.TemperatureCMB` and
+:class:`~astropy.cosmology.traits.DarkEnergyComponent` traits in a custom cosmology class:
 
 >>> import astropy.units as u
->>> from astropy.cosmology.traits import HubbleParameter, ScaleFactor, TemperatureCMB
+>>> from astropy.cosmology.traits import HubbleParameter, ScaleFactor, TemperatureCMB, DarkEnergyComponent
 >>> from astropy.cosmology import Cosmology
 >>>
->>> class CustomCosmology(Cosmology, HubbleParameter, ScaleFactor, TemperatureCMB):
+>>> class CustomCosmology(Cosmology, HubbleParameter, ScaleFactor, TemperatureCMB, DarkEnergyComponent):
 ...     def __init__(self, Om0, Ode0, H0=70, Tcmb0=2.725):
 ...         self.H0 = H0 << (u.km / u.s / u.Mpc)
 ...         self.Om0 = Om0
