### 2025-07-20 e2492d4
Example of a custom w(z)
