### 2025-07-20 6f0ab5e
Added missing numpy import in example
