### 2025-07-20 99e80d3
Matching return type to be np.float64
