### 2025-07-20 a5962fa
Dark Energy Component Trait Class (#18447)
* Initial commit of darkenergycomponent trait class
* Updating docs and changelog files
* Adding dark energy to public api

Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
