+++ astropy/wcs/src/wcslib_wrap.c
@@ -2964,7 +2964,7 @@ PyWcsprm_get_cunit(
     return NULL;
   }
 
-  if (self->unit_scaling != NULL) {
+  if (self->original_cunit != NULL) {
     return get_unit_list(
       "cunit", self->original_cunit, (Py_ssize_t)self->x.naxis, (PyObject*)self);
   } else {
@@ -2989,17 +2989,25 @@ PyWcsprm_set_cunit(
 
   note_change(self);
 
-  if (self->original_cunit != NULL) {
+  // Note that if the user explicitly changes the units we should now preserve
+  // those new units. The way the WCS class works normally when units are not
+  // preserved is that changing units on an existing WCS is not just a unit
+  // conversion but truly overriding units.
 
-    // Units have been set before, so special care is needed here. The class is being used
-    // in preserve_units=True mode, so this means that if the user explicitly changes the
-    // units we should now preserve those new units. The way the WCS class works normally
-    // when units are not preserved is that changing units on an existing WCS is not just
-    // a unit conversion but truly overriding units. So we need to do three things here:
-    // (a) update original_cunit to reflect the new units, (b) update cdelt/crval/cd in
-    // the underlying WCS so that they have the values they would have had if we had started
-    // off initializing the WCS with the new units, and (c) make sure the unit scaling is
-    // updated.
+  // There are three cases to consider now:
+  // (a) The user is not using preserve_units, in which case unit_scaling will
+  //     not be allocated. The units can be set directly on the WCS as normal.
+  // (b) The user is using preserve_units, but unit_scaling has not been allocated
+  //     as it has not been needed. In this case we can do as in (a) and if the user
+  //     then either calls set() explicitly or does a conversion, unit_scaling will
+  //     get allocated if needed.
+  // (c) The user is using preserve_units and the unit scaling has been allocated
+  //     and set. In this case, special case is needed
+
+  if (self->unit_scaling != NULL) {
+
+    // The class is being used in preserve_units=True mode and there has been at
+    // least one scale != 1 before.
 
     // Keep track of previous original units
     previous_original_cunit = malloc(self->x.naxis * sizeof(*previous_original_cunit));
@@ -3007,7 +3015,6 @@ PyWcsprm_set_cunit(
         PyErr_NoMemory();
         return -1;
     }
-
     for (int i = 0; i < self->x.naxis; ++i) {
         strncpy(previous_original_cunit[i], self->original_cunit[i], 72);
     }
@@ -3017,45 +3024,59 @@ PyWcsprm_set_cunit(
        (PyObject *)self, "cunit", value, (Py_ssize_t)self->x.naxis, self->original_cunit);
 
     // Check if there is a difference in scale between the units
-
     for (int i = 0; i < self->x.naxis; ++i) {
 
+        // For tabular axis, CDELTia and CRVALia relate to indices so we just
+        // copy the new unit and don't do any scaling.
+        if ((self->x.types[i] / 100) % 10 == 5) {
+            strncpy(self->x.cunit[i], self->original_cunit[i], 72);
+            continue;
+        }
+
         if (strcmp(previous_original_cunit[i], self->original_cunit[i]) != 0) {
-        status = wcsunits(previous_original_cunit[i], self->original_cunit[i], &scale, &offset, &power);
-        if (status != 0) {
-            free(previous_original_cunit);
-            PyErr_Format(
-            PyExc_ValueError,
-            "Test "
-            );
-            return -1;
+            status = wcsunits(previous_original_cunit[i], self->original_cunit[i], &scale, &offset, &power);
+            if (status != 0) {
+                free(previous_original_cunit);
+                PyErr_Format(
+                PyExc_ValueError,
[…]
