### 2025-07-21 edf2ade
Make it so that units can be changed on a WCS even if using preserve_units=True
