+++ astropy/wcs/tests/test_wcs.py
@@ -2280,6 +2280,48 @@ RADESYS = 'ICRS'               / Equatorial coordinate system
 
         assert_allclose(wcs_prog.all_pix2world(2, 3, 0), wcs_equiv.all_pix2world(2, 3, 0))
 
+    def test_change_cunit_original_deg(self):
+
+        # Similar to test_change_cunit but here we start off with a WCS in
+        # degrees so that there is no scaling going on, and then we start
+        # changing the values.
+
+        wcs_prog = wcs.WCS(naxis=2, preserve_units=True)
+        wcs_prog.wcs.ctype = "RA---TAN", "DEC--TAN"
+        wcs_prog.wcs.cunit = "deg", "deg"
+        wcs_prog.wcs.crval = 10, 20
+        wcs_prog.wcs.cdelt = 1, 2
+        wcs_prog.wcs.crpix = 1, 1
+
+        wcs_prog.wcs.set()
+
+        assert list(wcs_prog.wcs.cunit) == ["deg", "deg"]
+        assert_allclose(wcs_prog.wcs.crval, [10, 20])
+        assert_allclose(wcs_prog.wcs.cdelt, [1, 2])
+
+        assert wcs_prog.wcs._unit_scaling is None
+
+        wcs_prog.wcs.cunit = "arcsec", "arcsec"
+
+        wcs_prog.wcs.set()
+
+        assert_allclose(wcs_prog.wcs._unit_scaling, [1/3600, 1/3600])
+
+        assert list(wcs_prog.wcs.cunit) == ["arcsec", "arcsec"]
+        assert_allclose(wcs_prog.wcs.crval, [10, 20])
+        assert_allclose(wcs_prog.wcs.cdelt, [1, 2])
+
+        wcs_equiv = wcs.WCS(naxis=2)
+        wcs_equiv.wcs.ctype = "RA---TAN", "DEC--TAN"
+        wcs_equiv.wcs.cunit = "arcsec", "arcsec"
+        wcs_equiv.wcs.crval = 10, 20
+        wcs_equiv.wcs.cdelt = 1, 2
+        wcs_equiv.wcs.crpix = 1, 1
+        wcs_equiv.wcs.set()
+
+        assert_allclose(wcs_prog.all_pix2world(2, 3, 0)[0], wcs_equiv.all_pix2world(2, 3, 0)[0] * 3600)
+        assert_allclose(wcs_prog.all_pix2world(2, 3, 0)[1], wcs_equiv.all_pix2world(2, 3, 0)[1] * 3600)
+
     def test_str_and_repr(self):
         expected = (
             "WCS Keywords\n"
@@ -2352,12 +2394,16 @@ RADESYS = 'ICRS'               / Equatorial coordinate system
 
         assert simple_wcs.wcs._unit_scaling is None
 
-    def test_wcstab_cunit_conversion(self):
+    @pytest.mark.parametrize('mixed_with_spatial', (False, True))
+    def test_wcstab_cunit_conversion(self, mixed_with_spatial):
         # WCSLIB does not convert the units for coordinates that use lookup tables
         # with -TAB, but we should make sure that things still work properly
         # with preserve_units=True
 
-        flux = np.random.rand(100)
+        if mixed_with_spatial:
+            flux = np.random.rand(100, 4, 5)
+        else:
+            flux = np.random.rand(100)
 
         wav = np.linspace(4000, 5000, 100)
 
@@ -2368,7 +2414,7 @@ RADESYS = 'ICRS'               / Equatorial coordinate system
         tab_hdu.name = "WCS-TAB"
 
         header = primary_hdu.header
-        header["WCSAXES"] = 1
[…]
