### 2025-07-21 2cbb51b
Raise error if trying to modify  units in-place after set() has been called when using preserve_units=True, and also make sure we correctly raise errors related to incompatible units
