### 2025-07-21 847ec70
Raise error if trying to modify  units in-place after set() has been called when using preserve_units=True, and also make sure we correctly raise errors related to incompatible units
