### 2025-07-21 69323a3
Simplify implementation of set_cunit
