### 2025-07-21 1c7b0b1
compute max abs/rel differences on the full arrays
Fix #18413
