### 2025-07-21 e13204b
compute max abs/rel differences on the full arrays
Fix #18413
