### 2025-07-24 51e491c
WHL: test wheels against CPython 3.14 (GIL-enabled) (twin #18457)
