### 2025-07-24 c458536
Fix buffer handling and null detection to match original converter behavior
