### 2025-07-24 4684388
Fix buffer handling and null detection to match original converter behavior
