### 2025-07-24 693d366
Reduce duplication of binparse methods for Numeric types
