### 2025-07-25 27fd83c
TST: remove macOS + x86_64 combo from weekly cron
This is a direct backport of https://github.com/astropy/astropy/pull/18460 onto v7.1.x branch [ci skip]
