### 2025-07-25 3145615
Added default_preserve_units ScienceState
