### 2025-07-25 0bfcfb1
Move null handling to Cython parser implementations to address performance regression
