### 2025-07-25 98c4138
Move null handling to Cython parser implementations to address performance regression
