### 2025-07-25 f99fbb2
Remove unused `ErfaAstromInterpolator._get_c2i()`
The function has served no purpose since
6bf862a136f992c65f88a1d5f5c2fe6f485e3a40 (remove apci)
