### 2025-07-26 cb2fb44
Update docs/change entry with some notes on benchmark results
