### 2025-07-26 d03bd79
Add WCS.preserve_units property which indicates whether the WCS class is preserving the original units from the header or set programmatically, or whether some units may have been converted to SI or degrees
