### 2025-07-26 9a4dc77
Define ScienceState in wcs.py not top-level __init__.py
