### 2025-07-26 62c96d8
Explicitly list all imported objects from the astropy.wcs._wcs C extension, and remove fallbacks if extension can't be imported
