+++ astropy/wcs/wcs.py
@@ -52,13 +52,134 @@ from astropy.utils.exceptions import (
     AstropyWarning,
 )
 
-from . import _wcs, docstrings
+from . import docstrings
+from ._wcs import (
+    PRJ_CODES,
+    PRJ_CONIC,
+    PRJ_CONVENTIONAL,
+    PRJ_CYLINDRICAL,
+    PRJ_HEALPIX,
+    PRJ_POLYCONIC,
+    PRJ_PSEUDOCYLINDRICAL,
+    PRJ_PVN,
+    PRJ_QUADCUBE,
+    PRJ_ZENITHAL,
+    WCSCOMPARE_ANCILLARY,
+    WCSCOMPARE_CRPIX,
+    WCSCOMPARE_TILING,
+    WCSHDO_EFMT,
+    WCSHDO_P12,
+    WCSHDO_P13,
+    WCSHDO_P14,
+    WCSHDO_P15,
+    WCSHDO_P16,
+    WCSHDO_P17,
+    WCSHDR_ALLIMG,
+    WCSHDR_AUXIMG,
+    WCSHDR_BIMGARR,
+    WCSHDR_IMGHEAD,
+    WCSHDR_LONGKEY,
+    WCSHDR_PIXLIST,
+    WCSHDR_RADECSYS,
+    WCSHDR_VSOURCE,
+    WCSSUB_CELESTIAL,
+    WCSSUB_CUBEFACE,
+    WCSSUB_LATITUDE,
+    WCSSUB_LONGITUDE,
+    WCSSUB_SPECTRAL,
+    WCSSUB_STOKES,
+    WCSSUB_TIME,
+    Auxprm,
+    Celprm,
+    DistortionLookupTable,
+    InconsistentAxisTypesError,
+    InvalidCoordinateError,
+    InvalidPrjParametersError,
+    InvalidSubimageSpecificationError,
+    InvalidTabularParametersError,
+    InvalidTransformError,
+    NonseparableSubimageCoordinateSystemError,
+    NoSolutionError,
+    NoWcsKeywordsFoundError,
+    Prjprm,
+    SingularMatrixError,
+    Sip,
+    Tabprm,
+    WcsError,
+    WCSHDO_all,
+    WCSHDO_CNAMna,
+    WCSHDO_CRPXna,
+    WCSHDO_DOBSn,
+    WCSHDO_none,
+    WCSHDO_PVn_ma,
+    WCSHDO_safe,
+    WCSHDO_TPCn_ka,
+    WCSHDO_WCSNna,
+    WCSHDR_all,
+    WCSHDR_CD0i_0ja,
+    WCSHDR_CD00i00j,
+    WCSHDR_CNAMn,
+    WCSHDR_CROTAia,
+    WCSHDR_DOBSn,
+    WCSHDR_EPOCHa,
+    WCSHDR_none,
+    WCSHDR_PC0i_0ja,
+    WCSHDR_PC00i00j,
+    WCSHDR_PROJPn,
+    WCSHDR_PS0i_0ma,
+    WCSHDR_PV0i_0ma,
+    WCSHDR_reject,
+    WCSHDR_strict,
+    WCSHDR_VELREFa,
+    Wcsprm,
+    Wtbarr,
+    _sanity_check,
+    set_wtbarr_fitsio_callback,
+)
+from ._wcs import __version__ as __wcslib_version__
+from ._wcs import _Wcs as WCSBase
+from ._wcs import find_all_wcs as find_all_wcs_c
 
 # Mix-in class that provides the APE 14 API
[…]
