### 2025-07-26 59eb603
Explicitly list all imported objects from the astropy.wcs._wcs C extension, and remove fallbacks if extension can't be imported
