### 2025-07-26 5ec75eb
Rename _wcs.__version__ to _wcs.WCSLIB_VERSION
