+++ astropy/wcs/tests/test_wcs.py
@@ -35,14 +35,12 @@ from astropy.utils.exceptions import (
     AstropyWarning,
 )
 from astropy.utils.misc import NumpyRNGContext
-from astropy.wcs import _wcs
-
-_WCSLIB_VER = Version(_wcs.__version__)
+from astropy.wcs.wcs import WCSLIB_VERSION
 
 
 # NOTE: User can choose to use system wcslib instead of bundled.
 def ctx_for_v71_dateref_warnings():
-    if _WCSLIB_VER >= Version("7.1") and _WCSLIB_VER < Version("7.3"):
+    if Version(WCSLIB_VERSION) >= Version("7.1") and Version(WCSLIB_VERSION) < Version("7.3"):
         ctx = pytest.warns(
             wcs.FITSFixedWarning,
             match=(
@@ -124,7 +122,7 @@ class TestSpectra:
                 os.path.join("data", "spectra", filename), encoding="binary"
             )
             # finally run the test.
-            if _WCSLIB_VER >= Version("7.4"):
+            if Version(WCSLIB_VERSION) >= Version("7.4"):
                 ctx = pytest.warns(
                     wcs.FITSFixedWarning,
                     match=(
@@ -152,7 +150,7 @@ def test_fixes():
     ):
         wcs.WCS(header, translate_units="dhs")
 
-    if Version("7.4") <= _WCSLIB_VER < Version("7.6"):
+    if Version("7.4") <= Version(WCSLIB_VERSION) < Version("7.6"):
         assert len(w) == 3
         assert "'datfix' made the change 'Success'." in str(w.pop().message)
     else:
@@ -192,7 +190,7 @@ def test_pix2world():
         ww = wcs.WCS(filename)
 
     # might as well monitor for changing behavior
-    if Version("7.4") <= _WCSLIB_VER < Version("7.6"):
+    if Version("7.4") <= Version(WCSLIB_VERSION) < Version("7.6"):
         assert len(caught_warnings) == 2
     else:
         assert len(caught_warnings) == 1
@@ -253,10 +251,10 @@ def test_dict_init():
         "CDELT1": -0.1,
         "CDELT2": 0.1,
     }
-    if _WCSLIB_VER >= Version("7.1"):
+    if Version(WCSLIB_VERSION) >= Version("7.1"):
         hdr["DATEREF"] = "1858-11-17"
 
-    if _WCSLIB_VER >= Version("7.4"):
+    if Version(WCSLIB_VERSION) >= Version("7.4"):
         ctx = pytest.warns(
             wcs.wcs.FITSFixedWarning,
             match=r"'datfix' made the change 'Set MJDREF to 0\.000000 from DATEREF'\.",
@@ -369,7 +367,7 @@ def test_invalid_shape():
 
 def test_warning_about_defunct_keywords():
     header = get_pkg_data_contents("data/defunct_keywords.hdr", encoding="binary")
-    if Version("7.4") <= _WCSLIB_VER < Version("7.6"):
+    if Version("7.4") <= Version(WCSLIB_VERSION) < Version("7.6"):
         n_warn = 5
     else:
         n_warn = 4
@@ -405,14 +403,14 @@ def test_to_header_string():
     )
     # fmt: on
 
-    if _WCSLIB_VER >= Version("7.3"):
+    if Version(WCSLIB_VERSION) >= Version("7.3"):
         # fmt: off
         hdrstr += (
             "MJDREF  =                  0.0 / [d] MJD of fiducial time                       ",
         )
         # fmt: on
 
-    elif _WCSLIB_VER >= Version("7.1"):
+    elif Version(WCSLIB_VERSION) >= Version("7.1"):
[…]

[…]
