### 2025-07-28 3704a35
Added regression test
