### 2025-07-28 d5c1ccf
Remove two `ErfaAstromInterpolator` static methods
The removed static methods were only called by the `apco()` method and
their functionality is simpler to achieve using a local function defined
with `functools.partial()`.
