### 2025-07-28 344fe84
Remove redundant visibility assertions in tests
