### 2025-07-28 725b4cc
Split set_visible to only handle visibility
