### 2025-07-28 e3953d7
Refactor tests to access coords directly from ax
