### 2025-07-28 7b9f738
Split tests for set_visible and set_position
