+++ astropy/visualization/wcsaxes/tests/test_coordinate_helpers.py
@@ -279,22 +279,6 @@ def test_set_visible():
     assert ax.coords[0].get_ticklabel_visible() is True
     assert ax.coords[0].get_axislabel_visible() is True
 
-    ax.coords[0].set_visible("t")
-    assert ax.coords[0].get_ticks_position() == ["t"]
-    assert ax.coords[0].get_ticklabel_position() == ["t"]
-    assert ax.coords[0].get_axislabel_position() == ["t"]
-    assert ax.coords[0].get_ticks_visible() is True
-    assert ax.coords[0].get_ticklabel_visible() is True
-    assert ax.coords[0].get_axislabel_visible() is True
-
-    ax.coords[0].set_visible("bl")
-    assert ax.coords[0].get_ticks_position() == ["b", "l"]
-    assert ax.coords[0].get_ticklabel_position() == ["b", "l"]
-    assert ax.coords[0].get_axislabel_position() == ["b", "l"]
-    assert ax.coords[0].get_ticks_visible() is True
-    assert ax.coords[0].get_ticklabel_visible() is True
-    assert ax.coords[0].get_axislabel_visible() is True
-
 
 @pytest.mark.parametrize("invalid_input", [123, 123.45, None, [], {}])
 def test_set_visible_invalid_type(invalid_input):
@@ -305,6 +289,40 @@ def test_set_visible_invalid_type(invalid_input):
 
     with pytest.raises(TypeError):
         ax.coords[0].set_visible(invalid_input)
+
+
+def test_set_position():
+    fig = Figure()
+    canvas = FigureCanvasAgg(fig)
+    ax = WCSAxes(fig, [0.1, 0.1, 0.8, 0.8], wcs=WCS(MSX_HEADER))
+    fig.add_axes(ax)
+
+    ax.coords[0].set_position("t")
+    assert ax.coords[0].get_ticks_position() == ["t"]
+    assert ax.coords[0].get_ticklabel_position() == ["t"]
+    assert ax.coords[0].get_axislabel_position() == ["t"]
+    assert ax.coords[0].get_ticks_visible() is True
+    assert ax.coords[0].get_ticklabel_visible() is True
+    assert ax.coords[0].get_axislabel_visible() is True
+
+    ax.coords[0].set_position("bl")
+    assert ax.coords[0].get_ticks_position() == ["b", "l"]
+    assert ax.coords[0].get_ticklabel_position() == ["b", "l"]
+    assert ax.coords[0].get_axislabel_position() == ["b", "l"]
+    assert ax.coords[0].get_ticks_visible() is True
+    assert ax.coords[0].get_ticklabel_visible() is True
+    assert ax.coords[0].get_axislabel_visible() is True
+
+
+@pytest.mark.parametrize("invalid_input", [123, 123.45, None, [], {}])
+def test_set_position_invalid_type(invalid_input):
+    fig = Figure()
+    canvas = FigureCanvasAgg(fig)
+    ax = WCSAxes(fig, [0.1, 0.1, 0.8, 0.8], wcs=WCS(MSX_HEADER))
+    fig.add_axes(ax)
+
+    with pytest.raises(TypeError):
+        ax.coords[0].set_position(invalid_input)
 
 
 def test_set_position_invalid():
