### 2025-07-29 91f9711
C formatting
Co-authored-by: Mihai Cara <mcara@users.noreply.github.com>
