### 2025-07-29 60300af
Backport PR #16411: Avoid calling wcsset if the wcsprm is unchanged
