### 2025-07-29 0a1a47f
Return NamedTuples from coordinate match functions
The coordinate matching and searching functions and the corresponding
`SkyCoord` methods have so far returned tuples. Returning instances of
new `NamedTuple` subclasses allows accessing values in the returned
tuples via named attributes without breaking backwards compatibility.
