+++ astropy/coordinates/sky_coordinate.py
@@ -1162,20 +1162,13 @@ class SkyCoord(MaskableShapedLikeNDArray):
 
         Returns
         -------
-        idx : int array
-            Indices into ``catalogcoord`` to get the matched points for
-            each of this object's coordinates. Shape matches this
-            object.
-        sep2d : `~astropy.coordinates.Angle`
-            The on-sky separation between the closest match for each
-            element in this object in ``catalogcoord``. Shape matches
-            this object.
-        dist3d : `~astropy.units.Quantity` ['length']
-            The 3D distance between the closest match for each element
-            in this object in ``catalogcoord``. Shape matches this
-            object. Unless both this and ``catalogcoord`` have associated
-            distances, this quantity assumes that all sources are at a
-            distance of 1 (dimensionless).
+        CoordinateMatchResult
+            A `~typing.NamedTuple` with attributes representing for each
+            source in this |SkyCoord| the indices and angular and
+            physical separations of the match in ``catalogcoord``. If
+            either the |SkyCoord| or ``catalogcoord`` don't have
+            distances, the physical separation is the 3D distance on the
+            unit sphere, rather than a true distance.
 
         Notes
         -----
@@ -1230,18 +1223,10 @@ class SkyCoord(MaskableShapedLikeNDArray):
 
         Returns
         -------
-        idx : int array
-            Indices into ``catalogcoord`` to get the matched points for
-            each of this object's coordinates. Shape matches this
-            object.
-        sep2d : `~astropy.coordinates.Angle`
-            The on-sky separation between the closest match for each
-            element in this object in ``catalogcoord``. Shape matches
-            this object.
-        dist3d : `~astropy.units.Quantity` ['length']
-            The 3D distance between the closest match for each element
-            in this object in ``catalogcoord``. Shape matches this
-            object.
+        CoordinateMatchResult
+            A `~typing.NamedTuple` with attributes representing for each
+            source in this |SkyCoord| the indices and angular and physical
+            separations of the match in ``catalogcoord``.
 
         Notes
         -----
@@ -1292,20 +1277,13 @@ class SkyCoord(MaskableShapedLikeNDArray):
 
         Returns
         -------
-        idxsearcharound : int array
-            Indices into ``searcharoundcoords`` that match the
-            corresponding elements of ``idxself``. Shape matches
-            ``idxself``.
-        idxself : int array
-            Indices into ``self`` that match the
-            corresponding elements of ``idxsearcharound``. Shape matches
-            ``idxsearcharound``.
-        sep2d : `~astropy.coordinates.Angle`
-            The on-sky separation between the coordinates. Shape matches
-            ``idxsearcharound`` and ``idxself``.
-        dist3d : `~astropy.units.Quantity` ['length']
-            The 3D distance between the coordinates. Shape matches
-            ``idxsearcharound`` and ``idxself``.
+        CoordinateSearchResult
+            A `~typing.NamedTuple` with attributes representing the
+            indices of the elements of found pairs in the other set of
+            coordinates and this |SkyCoord| and angular and physical
+            separations of the pairs. If either set of sources lack
+            distances, the physical separation is the 3D distance on the
+            unit sphere, rather than a true distance.
 
         Notes
         -----
@@ -1352,20 +1330,11 @@ class SkyCoord(MaskableShapedLikeNDArray):
 
         Returns
         -------
-        idxsearcharound : int array
-            Indices into ``searcharoundcoords`` that match the
-            corresponding elements of ``idxself``. Shape matches
-            ``idxself``.
-        idxself : int array
-            Indices into ``self`` that match the
-            corresponding elements of ``idxsearcharound``. Shape matches
-            ``idxsearcharound``.
[…]
