### 2025-07-30 5d6c4d2
handle edge cases with invalid values
