### 2025-07-30 4e53373
Relax tolerance for HCOMPRESS_1 tests with quantization
Fixes test failures with HCOMPRESS_1 compression using SUBTRACTIVE_DITHER_2
quantization method by using appropriate tolerances for lossy compression.

Fixes #18468
