### 2025-07-30 3544eba
Simplify coordinates.concatenate_representations()
