### 2025-07-31 c7cc326
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
