### 2025-07-31 271005a
Backport PR #16411 on branch v7.1.x (Avoid calling wcsset if the wcsprm is unchanged) (#18472)
* Backport PR #16411: Avoid calling wcsset if the wcsprm is unchanged
