### 2025-08-01 e3c2428
Update IVOA UCD1+ controlled vocabulary from version 1.5 to 1.6
This update adds support for new atmospheric observation terms including
obs.atmos.wind, obs.atmos.humidity, obs.atmos.rain, obs.atmos.turbulence,
obs.atmos.turbulence.isoplanatic, and obs.atmos.water which are now
recognized when parsing UCDs with check_controlled_vocabulary=True.

Fixes #18452
