### 2025-08-01 b3b368f
Fix formatting issues in UCD test
- Remove trailing whitespace and blank line whitespace
- Fix PLR0133 by using variable instead of inline string comparison
- Ensure proper line formatting per ruff standards
