### 2025-08-01 168f273
Fix formatting and line ending issues
- Apply ruff formatting to test file
- Add missing newline to changelog file
- Ensure all pre-commit checks will pass
