### 2025-08-01 4e94f8e
Fix test_atmospheric_wind_ucd to properly handle UCD word types
Secondary words (obs.atmos.wind, obs.atmos.turbulence) can only be used
after primary words, while quantified/primary words can be used standalone.
The test now correctly separates these cases and tests secondary words
only in valid secondary positions.
