### 2025-08-01 31b9ddd
Apply ruff formatting to test_ucd.py
