### 2025-08-01 2b3e092
Update the RDB format description
