### 2025-08-01 fbc55f9
fix issue with heap when data is updated (fix #18479)
Once data is loaded we cannot use directly the heap data from disk,
since there is no way to know if the data is modified.
