+++ astropy/io/fits/fitsrec.py
@@ -1032,14 +1032,17 @@ class FITS_rec(np.recarray):
 
         return field
 
-    def _get_heap_data(self):
+    def _get_heap_data(self, try_from_disk=True):
         """
         Returns a pointer into the table's raw data to its heap (if present).
 
         This is returned as a numpy byte array.
         """
-        raw_data = self._get_raw_data()
-        if self._heapsize and raw_data is not None:
+        if (
+            try_from_disk
+            and self._heapsize
+            and (raw_data := self._get_raw_data()) is not None
+        ):
             # Read the heap from disk
             raw_data = raw_data.view(np.ubyte)
             heap_end = self._heapoffset + self._heapsize
