+++ astropy/io/fits/hdu/table.py
@@ -879,7 +879,7 @@ class BinTableHDU(_TableBaseHDU):
             # for the checksum calculation. We do this by padding the first few
             # bytes of the heap (if necessary), then calculating the checksum for
             # the rest of the heap data normally.
-            heap_data = data._get_heap_data()
+            heap_data = data._get_heap_data(try_from_disk=not self._data_loaded)
             if extra := self._theap % 4:
                 first_part = np.zeros(4, dtype=np.ubyte)
                 first_part[extra:] = heap_data[: 4 - extra]
@@ -927,7 +927,7 @@ class BinTableHDU(_TableBaseHDU):
 
             nbytes = data._gap
 
-            heap_data = data._get_heap_data()
+            heap_data = data._get_heap_data(try_from_disk=not self._data_loaded)
             if len(heap_data) > 0:
                 nbytes += len(heap_data)
                 fileobj.writearray(heap_data)
