### 2025-08-01 23f687b
use heap directly for compressed hdus
