### 2025-08-01 c6a3511
use heap directly for compressed hdus
