+++ astropy/utils/diff.py
@@ -204,7 +204,7 @@ def where_not_allclose(a, b, rtol=1e-5, atol=1e-8, return_maxdiff=False):
             max_absolute = max_relative = 0
         else:
             # remove all invalid values before computing max differences
-            relative = absolute / np.abs(b[finites])
+            relative = absolute / np.abs(b.data[finites])
             max_absolute = float(np.max(absolute))
             max_relative = np.max(relative)
         return indices, max_absolute, max_relative
