### 2025-08-01 f4703b4
mask invalid before computing differences
