### 2025-08-02 8fc2b7f
Update minimum required version of astropy-iers-data
