### 2025-08-02 264daae
Backport PR #18493: Update minimum required version of astropy-iers-data
