### 2025-08-03 dfaf6e2
Fix circular dependencies for `import astropy.io.misc.ecsv`
