### 2025-08-03 bd11099
Changes to allow frozen=True for ColumnECSV dataclass
