### 2025-08-03 c9603a2
Change ColumnECSV.dtype property to be numpy.dtype
