### 2025-08-04 6a22e39
Fix typing and documentation to resolve confusion about str_vals (always 1-d)
