+++ .pre-commit-config.yaml
@@ -64,7 +64,7 @@ repos:
           - tomli
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.2
+    rev: v0.12.7
     hooks:
       - id: ruff-check
         args: ["--fix", "--show-fixes"]
