### 2025-08-04 ab4fc54
Backport PR #18497: MAINT: minor cleanup of duplicate check
