### 2025-08-05 20ccfcd
Future proof skipping building free-threaded wheels
Direct backport of #18499 onto v7.1.x
