### 2025-08-06 651cf16
Add a test for formatting the liter
The new test demonstrates that "ℓ" is being used as the symbol for the
liter with some unit formatters, but according to the SI brochure
(https://www.bipm.org/documents/20126/41483022/SI-Brochure-9-EN.pdf) the
symbols for the liter are "l" or "L".
