+++ astropy/units/tests/test_format.py
@@ -1161,3 +1161,16 @@ def test_Fits_name_deprecation():
     ):
         from astropy.units.format import Fits
     assert Fits is u.format.FITS
+
+
+@pytest.mark.parametrize(
+    "format_spec",
+    [
+        "generic",
+        pytest.param(
+            "unicode", marks=pytest.mark.xfail(reason="regression test to reveal a bug")
+        ),
+    ],
+)
+def test_liter(format_spec):
+    assert format(u.liter, format_spec) == "l"
