### 2025-08-06 a0f41cc
Ensure the liter is always represented as `"l"`
Some unit formatters represented the liter as `"ℓ"` and others as `"l"`.
According to the SI brochure
(https://www.bipm.org/documents/20126/41483022/SI-Brochure-9-EN.pdf) the
symbol should be `"l"` or `"L"`, so the formatters that have so far used
`"ℓ"` now use `"l"` to be consistent with the others.
