### 2025-08-07 1ca00da
Add a test to reveal a bug in `rotation_matrix()`
According to the documentation the `angle` argument of
`coordinates.rotation_matrix()` can be an angle-like, but the new test
reveals that angle-like strings are nonetheless rejected.
