### 2025-08-07 f46ed2b
Fix a bug about `rotation_matrix()` `angle` types
According to the docstring of `rotation_matrix()` its `angle` argument
can be any angle-like value, but since
21ed682e57bfc56789b535bf4285190adfcd3a87 (Improve performance of
rotation_matrix for common angle units) and
223fae9b133da4e2dabba7a05cd290173a86f0e5 (Bug fix and suggested
improvements) it has rejected some angle-like types, e.g. angle-like
strings. The new code fixes the bug and retains a shortcut for
`Quantity` (and `Angle`) instances.
