### 2025-08-08 2168aca
Read CDS tables where the data is split in separate files
Part of the CDS specs is that data files can be split to keep each ASCII file below 10 MB. There are few (possibly only the Tycho 2 catalog) such cases in CDS, but we want to be complete.
The implementation is in the CDS reader, looking for the very specific case where a ReadMe file is given, and the catalog file is not found with the expected name. In that and only that case, the code checks if split data files exist.

closes #4121
