### 2025-08-08 df9cc37
Improve error message when column formatting fails (#17631)
* Improve error message when column formatting fails

Add additional details to the error message when setting the format of a table column fails.
The message is not complete in the sense that it does not point the user to all possible settings (e.g. old-style formatting is also allowed), but it should set expectations right and make clear that the `.format` is just for printing and will not change the dtype of a column (see #8071).

* Add changelog

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see htt
[…]
