+++ astropy/io/misc/ecsv.py
@@ -67,7 +67,7 @@ from collections.abc import Iterable
 from contextlib import ExitStack
 from dataclasses import dataclass
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Final, Literal, NamedTuple
+from typing import TYPE_CHECKING, Any, Final, NamedTuple
 
 import numpy as np
 import numpy.typing as npt
@@ -78,6 +78,9 @@ if TYPE_CHECKING:
 __all__ = [
     "ColumnECSV",
     "ECSVEngine",
+    "ECSVEngineIoAscii",
+    "ECSVEnginePandas",
+    "ECSVEnginePyArrow",
     "ECSVHeader",
     "read_ecsv",
     "register_ecsv_table",
@@ -327,11 +330,7 @@ class ECSVEnginePyArrow(ECSVEngine):
         header: ECSVHeader,
         null_values: list[str],
     ) -> dict[str, Any]:
-        """
-        Generate a dictionary of keyword arguments for data parsing.
-
-        See base method for details.
-        """
+        # See base method for details.
         kw = {}
         kw["null_values"] = null_values
         kw["header_start"] = header.n_header
@@ -345,9 +344,10 @@ class ECSVEngineIoAscii(ECSVEngine):
     name = "io.ascii"
     format = "ascii.csv"
 
-    def convert_np_type(self, np_type: str) -> str:
-        """Convert the np_type string to a numpy dtype type."""
-        # Output compatible with io.ascii `converters` option.
+    def convert_np_type(self, np_type: str) -> np.generic:
+        # Convert the np_type string to a numpy dtype type like np.int32, np.float64,
+        # etc. This output is compatible with io.ascii `converters` option where is gets
+        # used.
         return np.dtype(np_type).type
 
     def get_data_kwargs(
@@ -355,7 +355,6 @@ class ECSVEngineIoAscii(ECSVEngine):
         header: ECSVHeader,
         null_values: list[str],
     ) -> dict[str, Any]:
-        """Get the keyword arguments for reading the data."""
         kw = {}
         kw["fill_values"] = get_null_values_per_column(
             header.cols, header.table_meta, null_values
@@ -376,7 +375,7 @@ class ECSVEnginePandas(ECSVEngine):
     format = "pandas.csv"
 
     def convert_np_type(self, np_type: str) -> np.dtype:
-        """Convert the np_type to a pandas dtype will support for nullable types."""
+        # Convert the np_type to a pandas dtype will support for nullable types.
         import pandas as pd
 
         dtype = np.dtype(np_type)
@@ -416,7 +415,7 @@ class ECSVEnginePandas(ECSVEngine):
 
 
 def is_numpy_dtype(np_type: str) -> bool:
-    """Check if the given dtype is a valid numpy dtype."""
+    # Check if the given dtype is a valid numpy dtype.
     try:
         np.dtype(np_type)
     except Exception:
@@ -685,7 +684,7 @@ def read_data(
     header: ECSVHeader,
     null_values: list[str],
     encoding: str = "utf-8",
-    engine_name: Literal["pyarrow", "io.ascii", "pandas"] = "io.ascii",
+    engine_name: str = "io.ascii",
 ) -> "Table":
     """
     Read the data from an ECSV table using the specified engine.
@@ -703,8 +702,9 @@ def read_data(
         List of string values to interpret as null/missing values in the data.
     encoding : str, optional
         The encoding to use when reading the file. Default is "utf-8".
-    engine_name: Literal["pyarrow", "io.ascii", "pandas"], optional
+    engine_name: str, optional
         The backend engine to use for reading the data. Default is "io.ascii".
+        Built-in options are "pyarrow", "pandas", and "io.ascii".
 
[…]
