### 2025-08-08 808039d
Add more checks to ensure new code path is not triggered accidentially
tighten up regular expression

reorder imports
