### 2025-08-08 752e8d7
review updated and  auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
