### 2025-08-08 9aa1e90
review updated and  auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
