### 2025-08-08 6f7b62a
Find metadata in CDS readme even for gzipped files
CDS/Vizier serves large datafiles in gzipped for by default, but the ".gz" ending is not included in the filename in the ReadMe file and thus the reader will not find the metadata and fail.
This PR simply strips off any ".gz" ending before searching the ReadMe.

closes #6549
