+++ astropy/io/ascii/cds.py
@@ -108,7 +108,9 @@ class CdsHeader(core.BaseHeader):
                         # Iterate on names to find if one matches the tablename
                         # including wildcards.
                         for pattern in names:
-                            if fnmatch.fnmatch(self.data.table_name, pattern):
+                            if fnmatch.fnmatch(
+                                self.data.table_name.removesuffix(".gz"), pattern
+                            ):
                                 in_header = True
                                 lines.append(line)
                                 break
