### 2025-08-08 2231af8
Add new ECSV reader allowing different engines for CSV reader (#18267)
* Initial implementation of pyarrow engine for ECSV

* Working for simple test cases

* Passing minimally mod'd io.ascii ECSV tests

* Allow for different null values and compressed files

* Change ascii.read to Table.read where possible

* Mostly working refactor of type handling

* Refactor passing tests

* Baseline update io.ascii tests for both pyarrow and io.ascii engines (not passing)

* All tests except float16 passing

* Mostly complete refactor

* Still more refactoring

* Remove unnecessary default setting for dty
[…]
