+++ astropy/io/misc/ecsv.py
@@ -1,0 +1,1254 @@
+"""
+ECSV Engine Module
+--------------------------
+This module provides functionality for reading and writing Enhanced Character Separated
+Values (ECSV) files using various backends, including:
+- `PyArrow <https://arrow.apache.org/docs/python/>`_
+- `pandas <https://pandas.pydata.org/>`_
+- `Astropy's ASCII engine <https://docs.astropy.org/en/stable/io/ascii/>`_
+
+ECSV is a human-readable, YAML-encoded table format used in the Astropy
+ecosystem for storing tables with metadata, units, and complex data types.
+
+Key Features
+------------
+- Defines data structures for representing ECSV column and header metadata.
+- Implements multiple ECSV reader engines, supporting PyArrow, pandas, and Astropy's
+  ASCII CSV readers.
+- Handles conversion between ECSV datatypes and numpy/pandas/pyarrow types, including
+  support for JSON-encoded columns, multidimensional arrays, and masked data.
+- Provides robust parsing of ECSV headers and data, including support for compressed
+  files and in-memory file-like objects.
+- Ensures compatibility with legacy ECSV files and provides liberal error handling for
+  unknown datatypes.
+- Integrates with Astropy's Unified I/O registry for seamless reading and writing of
+  ECSV files.
+
+Main Classes and Functions
+--------------------------
+- ``ColumnECSV``: Represents the attributes of a column as described in the ECSV header.
+- ``ECSVHeader``: Encapsulates the parsed ECSV header, including column definitions and
+  table metadata.
+- ``ECSVEngine`` and subclasses: Abstract base class and concrete implementations for
+  different CSV parsing engines.
+- ``read_ecsv``: Reads an ECSV file and returns an Astropy ``Table`` object, handling all
+  necessary conversions and metadata.
+- ``write_ecsv``: Writes an Astropy ``Table`` to an ECSV file.
+- ``register_pyarrow_ecsv_table``: Registers the PyArrow ECSV reader/writer with
+  Astropy's I/O registry.
+
+Usage
+-----
+This module is intended for internal use within Astropy and for advanced users who need
+fine-grained control over ECSV parsing and engine selection. For most users, reading and
+writing ECSV files can be accomplished via the high-level ``Table.read`` and
+``Table.write`` interfaces, e.g.:
+```
+Table.read(filename, format="ecsv", engine="pyarrow.csv")
+```
+
+Dependencies
+------------
+- numpy
+- astropy.table
+- pyarrow (optional, for PyArrow engine)
+- pandas (optional, for pandas engine)
+"""
+
+import abc
+import collections
+import functools
+import io
+import json
+import os
+import re
+import warnings
+from collections.abc import Iterable
+from contextlib import ExitStack
+from dataclasses import dataclass
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Final, NamedTuple
+
+import numpy as np
+import numpy.typing as npt
+
+if TYPE_CHECKING:
+    from astropy.table import SerializedColumn, Table
+
+__all__ = [
+    "ColumnECSV",
+    "ECSVEngine",
+    "ECSVEngineIoAscii",
+    "ECSVEnginePandas",
+    "ECSVEnginePyArrow",
+    "ECSVHeader",
+    "read_ecsv",
+    "register_ecsv_table",
+    "write_ecsv",
+]
+
+ECSVEngines: Final[dict[str, "ECSVEngine"]] = {}
+
+
+class DerivedColumnProperties(NamedTuple):
+    """Named tuple for derived properties of a ECSV column specification.
+
+    Attributes
+    ----------
+    csv_np_type : str
+        Numpy type string for the CSV column data, e.g. "int64", "float32", "str".
[…]
