+++ astropy/coordinates/angles/utils.py
@@ -24,6 +24,8 @@ from astropy.coordinates.representation import (
 )
 from astropy.utils.compat import COPY_IF_NEEDED
 
+from .core import Angle
+
 _TWOPI = 2 * np.pi
 
 
@@ -84,8 +86,6 @@ def position_angle(lon1, lat1, lon2, lat2):
         following the appropriate `numpy` broadcasting rules.
 
     """
-    from .core import Angle
-
     deltalon = lon2 - lon1
     colat = np.cos(lat2)
 
@@ -114,8 +114,6 @@ def offset_by(lon, lat, posang, distance):
         these will contain arrays following the appropriate `numpy` broadcasting rules.
         0 <= lon < 2pi.
     """
-    from .core import Angle
-
     # Calculations are done using the spherical trigonometry sine and cosine rules
     # of the triangle A at North Pole,   B at starting point,   C at final point
     # with angles     A (change in lon), B (posang),            C (not used, but negative reciprocal posang)
