### 2025-08-11 24715f4
Address PR feedback: remove has_colon parameter and add phys.temperature.dew
