### 2025-08-11 ef574cb
Fix trailing whitespace in changelog
