### 2025-08-11 4485236
Change the quotes in the repr of units from " to '
The repr of Python strings uses `'`, not `"`, so for consistency the
repr of `astropy` units should also use `'`. This change does not affect
the representation of `Quantity` instances.
