+++ astropy/nddata/tests/test_nddata.py
@@ -520,8 +520,8 @@ def test_nddata_repr_dask():
     s = repr(arr)
     # just check repr equality for dask arrays, not round-tripping:
     assert s in (
-        'NDData(\n  data=dask.array<arange, shape=(3,), dtype=int64, chunksize=(3,), chunktype=numpy.ndarray>,\n  unit=Unit("km")\n)',
-        'NDData(\n  data=dask.array<arange, shape=(3,), dtype=int32, chunksize=(3,), chunktype=numpy.ndarray>,\n  unit=Unit("km")\n)',
+        "NDData(\n  data=dask.array<arange, shape=(3,), dtype=int64, chunksize=(3,), chunktype=numpy.ndarray>,\n  unit=Unit('km')\n)",
+        "NDData(\n  data=dask.array<arange, shape=(3,), dtype=int32, chunksize=(3,), chunktype=numpy.ndarray>,\n  unit=Unit('km')\n)",
     )
 
 
