### 2025-08-12 05c2c7f
Initial refactor
