+++ astropy/table/dataframes.py
@@ -1,0 +1,601 @@
+"""
+DataFrame conversion utilities for Astropy Tables.
+
+This module provides a class-based implementation for converting between
+Astropy Tables and various DataFrame formats (pandas, polars, etc) via
+the narwhals library.
+"""
+
+import warnings
+from collections import OrderedDict
+from collections.abc import Mapping
+
+import numpy as np
+
+from .column import Column, MaskedColumn
+
+
+def _encode_mixins(tbl):
+    """Encode mixin columns to basic columns for DataFrame compatibility."""
+    from astropy.time import TimeBase, TimeDelta
+
+    from . import serialize
+    from .column import col_copy
+
+    time_cols = [col for col in tbl.itercols() if isinstance(col, TimeBase)]
+    if time_cols:
+        new_cols = []
+        for col in tbl.itercols():
+            new_col = col_copy(col, copy_indices=False) if col.info.indices else col
+            new_cols.append(new_col)
+        tbl = tbl.__class__(new_cols, copy=False)
+
+        for col in tbl.itercols():
+            col.info.indices.clear()
+
+        for col in time_cols:
+            if isinstance(col, TimeDelta):
+                new_col = (col.sec * 1e9).astype("timedelta64[ns]")
+                nat = np.timedelta64("NaT")
+            else:
+                new_col = col.datetime64.copy()
+                nat = np.datetime64("NaT")
+            if col.masked:
+                new_col[col.mask] = nat
+            tbl[col.info.name] = new_col
+
+    # Convert the table to one with no mixins, only Column objects.
+    encode_tbl = serialize.represent_mixins_as_columns(tbl)
+    return encode_tbl
+
+
+class DataFrameConverter:
+    """
+    A class for converting between Astropy Tables and various DataFrame formats.
+
+    This class provides backend-specific handling for DataFrame conversions,
+    including validation of input columns and dependency checking.
+    """
+
+    def __init__(self, backend=None):
+        """
+        Initialize the DataFrameConverter.
+
+        Parameters
+        ----------
+        backend : str, module, or narwhals.Implementation, optional
+            The backend to use for conversions. If None, will be determined
+            at conversion time.
+        """
+        self.backend = backend
+        self._backend_impl = None
+
+    def _get_backend_impl(self, backend=None):
+        """Get the narwhals backend implementation."""
+        if backend is None:
+            backend = self.backend
+
+        if backend is None:
+            raise ValueError("No backend specified")
+
+        try:
+            import narwhals as nw
+        except ImportError:
+            raise ImportError(
+                "The narwhals library is required for DataFrame conversion. "
+                "Install with: pip install narwhals"
+            )
+
+        return nw.Implementation.from_backend(backend)
+
+    def _validate_columns_for_backend(self, table, backend_impl):
+        """
+        Validate that table columns are compatible with the target backend.
+
+        Parameters
+        ----------
+        table : Table
+            The table to validate
+        backend_impl : narwhals.Implementation
+            The backend implementation
+
+        Raises
+        ------
+        ValueError
+            If columns are not compatible with the backend
+        """
+        # Check for multidimensional columns
+        badcols = [name for name, col in table.columns.items() if len(col.shape) > 1]
+        if badcols:
+            raise ValueError(
+                f"Cannot convert a table with multidimensional columns to a "
+                f"DataFrame. Offending columns are: {badcols}\n"
+                f"One can filter out such columns using:\n"
+                f"names = [name for name in tbl.colnames if len(tbl[name].shape) <= 1]\n"
+                f"tbl[names].to_pandas(...)"
+            )
+
+    def _handle_index_argument(self, table, index, backend_impl):
+        """
+        Process the index argument for DataFrame conversion.
+
+        Parameters
[…]
