### 2025-08-12 22ab8fb
Initial refactor
