### 2025-08-12 dfca307
Initial refactor
