+++ astropy/table/table.py
@@ -4004,36 +4004,9 @@ class Table:
         return groups.table_group_by(self, keys)
 
     def _encode_mixins(tbl):
-        """Encode mixin columns to basic columns for Pandas/Polars compatibility."""
-        from astropy.time import TimeBase, TimeDelta
-
-        from . import serialize
-
-        time_cols = [col for col in tbl.itercols() if isinstance(col, TimeBase)]
-        if time_cols:
-            new_cols = []
-            for col in tbl.itercols():
-                new_col = col_copy(col, copy_indices=False) if col.info.indices else col
-                new_cols.append(new_col)
-            tbl = tbl.__class__(new_cols, copy=False)
-
-            for col in tbl.itercols():
-                col.info.indices.clear()
-
-            for col in time_cols:
-                if isinstance(col, TimeDelta):
-                    new_col = (col.sec * 1e9).astype("timedelta64[ns]")
-                    nat = np.timedelta64("NaT")
-                else:
-                    new_col = col.datetime64.copy()
-                    nat = np.datetime64("NaT")
-                if col.masked:
-                    new_col[col.mask] = nat
-                tbl[col.info.name] = new_col
-
-        # Convert the table to one with no mixins, only Column objects.
-        encode_tbl = serialize.represent_mixins_as_columns(tbl)
-        return encode_tbl
+        """Encode mixin columns to basic columns for DataFrame compatibility."""
+        from .dataframes import _encode_mixins
+        return _encode_mixins(tbl)
 
     def to_pandas(self, index=None, use_nullable_int=True):
         """
@@ -4102,85 +4075,8 @@ class Table:
           2002-01-01  2.0    6.0     8.0 0 days 00:03:20
 
         """
-        from pandas import DataFrame, Series
-
-        # Handle index argument
-        has_single_pk = self.primary_key and len(self.primary_key) == 1
-        if index is not False:
-            if index is True or index is None:
-                if has_single_pk:
-                    index = self.primary_key[0]
-                else:
-                    if index is True:
-                        raise ValueError(
-                            "index=True requires a single-column primary key."
-                        )
-                    else:
-                        index = False
-            elif isinstance(index, str):
-                if index not in self.colnames:
-                    raise ValueError(f"'{index}' is not in the table columns.")
-            else:
-                raise ValueError(
-                    "index must be None, False, True, or a valid column name."
-                )
-
-        tbl = self._encode_mixins()
-
-        badcols = [name for name, col in self.columns.items() if len(col.shape) > 1]
-        if badcols:
-            raise ValueError(
-                f"Cannot convert a table with multidimensional columns to a "
-                f"pandas DataFrame. Offending columns are: {badcols}\n"
-                f"One can filter out such columns using:\n"
-                f"names = [name for name in tbl.colnames if len(tbl[name].shape) <= 1]\n"
-                f"tbl[names].to_pandas(...)"
-            )
-
-        out = OrderedDict()
-
-        for name, column in tbl.columns.items():
-            if getattr(column.dtype, "isnative", True):
-                out[name] = column
-            else:
-                out[name] = column.data.byteswap().view(column.dtype.newbyteorder("="))
-
-            if isinstance(column, MaskedColumn) and np.any(column.mask):
-                if column.dtype.kind in ["i", "u"]:
-                    pd_dtype = column.dtype.name
-                    if use_nullable_int:
-                        # Convert int64 to Int64, uint32 to UInt32, etc for nullable types
-                        pd_dtype = pd_dtype.replace("i", "I").replace("u", "U")
-                    else:
-                        raise ValueError(
-                            "Cannot convert masked integer columns to DataFrame without using nullable integers."
-                            f"Set use_nullable_int=True or remove the offending column: {name}."
[…]
