### 2025-08-12 cb565cb
New docs and refactored tests
