+++ astropy/table/tests/test_df.py
@@ -1,0 +1,517 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+import numpy as np
+import pytest
+from numpy.testing import assert_allclose
+
+from astropy import table
+from astropy import units as u
+from astropy.table import MaskedColumn, Table
+from astropy.time import Time, TimeDelta
+from astropy.utils.compat.optional_deps import HAS_NARWHALS, HAS_PANDAS, HAS_POLARS
+
+from .conftest import MIXIN_COLS
+
+
+@pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")
+class TestPandasLegacy:
+    """Test the legacy pandas-specific to_pandas/from_pandas methods."""
+
+    def test_simple(self):
+        t = table.Table()
+
+        for endian in ["<", ">", "="]:
+            for kind in ["f", "i"]:
+                for byte in ["2", "4", "8"]:
+                    dtype = np.dtype(endian + kind + byte)
+                    x = np.array([1, 2, 3], dtype=dtype)
+                    t[endian + kind + byte] = x.view(x.dtype.newbyteorder(endian))
+
+        t["u"] = ["a", "b", "c"]
+        t["s"] = ["a", "b", "c"]
+
+        d = t.to_pandas()
+
+        for column in t.columns:
+            if column == "u":
+                assert np.all(t["u"] == np.array(["a", "b", "c"]))
+                assert d[column].dtype == np.dtype("O")  # upstream feature of pandas
+            elif column == "s":
+                assert np.all(t["s"] == np.array(["a", "b", "c"]))
+                assert d[column].dtype == np.dtype("O")  # upstream feature of pandas
+            else:
+                # We should be able to compare exact values here
+                assert np.all(t[column] == d[column])
+                if t[column].dtype.isnative:
+                    assert d[column].dtype == t[column].dtype
+                else:
+                    assert d[column].dtype == t[column].dtype.newbyteorder()
+
+        # Regression test for astropy/astropy#1156 - the following code gave a
+        # ValueError: Big-endian buffer not supported on little-endian
+        # compiler. We now automatically swap the endian-ness to native order
+        # upon adding the arrays to the data frame.
+        # Explicitly testing little/big/native endian separately -
+        # regression for a case in astropy/astropy#11286 not caught by #3729.
+        d[["<i4", ">i4"]]
+        d[["<f4", ">f4"]]
+
+        t2 = table.Table.from_pandas(d)
+
+        for column in t.columns:
+            if column in ("u", "s"):
+                assert np.all(t[column] == t2[column])
+            else:
+                assert_allclose(t[column], t2[column])
+            if t[column].dtype.isnative:
+                assert t[column].dtype == t2[column].dtype
+            else:
+                assert t[column].dtype.newbyteorder() == t2[column].dtype
+
+    @pytest.mark.parametrize("use_IndexedTable", [False, True])
+    def test_to_pandas_index(self, use_IndexedTable):
+        """Test to_pandas() with different indexing options.
+
+        This also tests the fix for #12014. The exception seen there is
+        reproduced here without the fix.
+        """
+        import pandas as pd
+
+        class IndexedTable(table.QTable):
+            """Always index the first column"""
+
+            def __init__(self, *args, **kwargs):
+                super().__init__(*args, **kwargs)
+                self.add_index(self.colnames[0])
+
[…]
