+++ docs/table/df_narwhals.rst
@@ -1,178 +1,0 @@
-.. doctest-skip-all
-
-.. _df_narwhals:
-
-Interfacing with Pandas/Polars DataFrames via Narwhals
-******************************************************
-
-`Narwhals <https://narwhals-dev.github.io/narwhals/>`_ provides a unified translation layer for DataFrame-like objects in Python, including `pandas <https://pandas.pydata.org/>`_, `polars <https://pola.rs/>`__, and others. This enables seamless interoperability between :class:`~astropy.table.Table` and DataFrame objects using generic methods: :meth:`~astropy.table.Table.to_df` and :meth:`~astropy.table.Table.from_df`.
-
-Narwhals supports multiple DataFrame libraries. The following is supported within Astropy:
-
-* Importing from any Narwhals compatible DataFrame, such as `pandas`, ``polars``, ``pyarrow``, ``duckDB``, and others.
-* Exporting to Narwhals compatible DataFrames with eager support, such as `pandas`, ``polars``, ``pyarrow``, ``modin``, or ``cudf``.
-
-The following examples and caveats will focus on common DataFrame libraries, primarily `pandas`.
-
-Basic Example
--------------
-
-.. EXAMPLE START: Interfacing Tables with DataFrames via Narwhals
-
-To demonstrate, we can create a minimal table::
-
-    >>> from astropy.table import Table
-    >>> t = Table()
-    >>> t['a'] = [1, 2, 3, 4]
-    >>> t['b'] = ['a', 'b', 'c', 'd']
-
-Which we can then convert to a pandas DataFrame::
-
-    >>> df = t.to_df("pandas")
-    >>> df
-       a  b
-    0  1  a
-    1  2  b
-    2  3  c
-    3  4  d
-    >>> type(df)
-    <class 'pandas.core.frame.DataFrame'>
-
-To convert to a polars DataFrame, we can specify the polars backend. We can also pass the backend as a module directly::
-
-    >>> import polars as pl
-    >>> df_polars = t.to_df(pl)
-    >>> type(df_polars)
-    <class 'polars.dataframe.frame.DataFrame'>
-
-It is also possible to create a table from a DataFrame::
-
-    >>> t2 = Table.from_df(df)
-    >>> t2
-    <Table length=4>
-      a      b
-    int64 string8
-    ----- -------
-        1       a
-        2       b
-        3       c
-        4       d
-
-.. EXAMPLE END
-
-Details and Limitations
------------------------
-
-Conversions to and from DataFrames are subject to several caveats:
-
-* Multidimensional columns have varied support across DataFrame libraries. Pandas does not support multidimensional columns natively, while Polars can support Nd arrays. PyArrow can support 1D arrays, but higher dimensions are not supported. Exercise caution when converting tables with multidimensional columns.
-* Mixed-type columns, i.e. object dtype columns, have varied support across DataFrame libraries. Where possible, we attempt to preserve the original data type, but this may not always be possible. For example, Pandas will preserve object-typed columns while Polars and PyArrow may fail to import the data at all.
-* Masked tables are partially supported. DataFrames use sentinel values (e.g., `numpy.nan` or `None`) for missing data, while Astropy preserves the original value under the mask. Masked values may be lost or replaced during conversion.
[…]
