+++ docs/table/dataframes.rst
@@ -1,0 +1,286 @@
+.. doctest-skip-all
+
+.. _df_narwhals:
+
+Interfacing with DataFrames
+***************************
+
+The :class:`~astropy.table.Table` class provides comprehensive support for interfacing with DataFrame libraries through two complementary approaches:
+
+1. **Legacy pandas-specific methods**: :meth:`~astropy.table.Table.to_pandas` and :meth:`~astropy.table.Table.from_pandas` for direct pandas integration
+2. **Generic narwhals-based methods**: :meth:`~astropy.table.Table.to_df` and :meth:`~astropy.table.Table.from_df` for multi-backend DataFrame support
+
+The narwhals-based approach uses `Narwhals <https://narwhals-dev.github.io/narwhals/>`_ as a unified translation layer, enabling seamless interoperability with multiple DataFrame libraries including `pandas <https://pandas.pydata.org/>`_, `polars <https://pola.rs/>`__, `pyarrow <https://arrow.apache.org/docs/python/>`__, and others.
+
+Pandas-Specific Methods
+=======================
+
+For pandas users, Astropy provides dedicated methods that offer the most direct and feature-complete integration with pandas DataFrames.
+
+Basic Pandas Example
+--------------------
+
+.. EXAMPLE START: Using Pandas-Specific Methods
+
+To demonstrate, we can create a minimal table::
+
+    >>> from astropy.table import Table
+    >>> t = Table()
+    >>> t['a'] = [1, 2, 3, 4]
+    >>> t['b'] = ['a', 'b', 'c', 'd']
+
+Convert to a pandas DataFrame using the pandas-specific method::
+
+    >>> df = t.to_pandas()
+    >>> df
+       a  b
+    0  1  a
+    1  2  b
+    2  3  c
+    3  4  d
+    >>> type(df)
+    <class 'pandas.core.frame.DataFrame'>
+
+Create a table from a pandas DataFrame::
+
+    >>> t2 = Table.from_pandas(df)
+    >>> t2
+    <Table length=4>
+      a      b
+    int64 string8
+    ----- -------
+        1       a
+        2       b
+        3       c
+        4       d
+
+.. EXAMPLE END
+
+Pandas Index Support
+--------------------
+
+The pandas-specific methods provide full support for DataFrame indexing, which is a unique pandas feature::
+
+    >>> from astropy.time import Time
+    >>> tm = Time([1998, 2002], format="jyear")
+    >>> x = [1, 2]
+    >>> t = Table([tm, x], names=["tm", "x"])
+    
+    # Use a column as the DataFrame index
+    >>> df = t.to_pandas(index="tm")
+    >>> df.index.name
+    'tm'
+    
+    # Convert back including the index as a column
+    >>> t_back = Table.from_pandas(df, index=True)
+    >>> t_back.colnames
+    ['tm', 'x']
+
+Generic Multi-Backend Methods
+=============================
+
+For users working with multiple DataFrame libraries or seeking broader compatibility, the generic methods provide a unified interface through narwhals.
+
+Supported Backends
+------------------
+
+The generic methods support DataFrame libraries with eager execution, including:
+
+* **pandas** - The most feature-complete backend
+* **polars** - High-performance DataFrame library with different handling of multidimensional data
+* **pyarrow** - In-memory columnar format with good performance
+* **modin** - Distributed pandas-compatible DataFrames
+* **cudf** - GPU-accelerated DataFrames
+
+.. note::
[…]
