### 2025-08-12 f00979a
New docs and refactored tests
