+++ astropy/table/tests/test_table.py
@@ -13,7 +13,7 @@ from io import StringIO
 
 import numpy as np
 import pytest
-from numpy.testing import assert_allclose, assert_array_equal
+from numpy.testing import assert_array_equal
 
 from astropy import table
 from astropy import units as u
@@ -28,14 +28,13 @@ from astropy.table import (
     TableReplaceWarning,
 )
 from astropy.tests.helper import assert_follows_unicode_guidelines
-from astropy.time import Time, TimeDelta
+from astropy.time import Time
 from astropy.utils.compat import NUMPY_LT_1_25
-from astropy.utils.compat.optional_deps import HAS_NARWHALS, HAS_PANDAS, HAS_POLARS
 from astropy.utils.data import get_pkg_data_filename
 from astropy.utils.exceptions import AstropyDeprecationWarning, AstropyUserWarning
 from astropy.utils.metadata.tests.test_metadata import MetaBaseTest
 
-from .conftest import MIXIN_COLS, MaskedTable
+from .conftest import MaskedTable
 
 
 @pytest.fixture
@@ -2061,525 +2060,6 @@ def test_table_init_from_degenerate_arrays(table_types):
 
     t = table_types.Table(np.array([1, 2, 3]))
     assert len(t.columns) == 3
-
-
-@pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")
-@pytest.mark.skipif(not HAS_NARWHALS, reason="requires narwhals")
-class TestPandas:
-    def test_simple(self):
-        t = table.Table()
-
-        for endian in ["<", ">", "="]:
-            for kind in ["f", "i"]:
-                for byte in ["2", "4", "8"]:
-                    dtype = np.dtype(endian + kind + byte)
-                    x = np.array([1, 2, 3], dtype=dtype)
-                    t[endian + kind + byte] = x.view(x.dtype.newbyteorder(endian))
-
-        t["u"] = ["a", "b", "c"]
-        t["s"] = ["a", "b", "c"]
-
-        d = t.to_pandas()
-
-        for column in t.columns:
-            if column == "u":
-                assert np.all(t["u"] == np.array(["a", "b", "c"]))
-                assert d[column].dtype == np.dtype("O")  # upstream feature of pandas
-            elif column == "s":
-                assert np.all(t["s"] == np.array(["a", "b", "c"]))
-                assert d[column].dtype == np.dtype("O")  # upstream feature of pandas
-            else:
-                # We should be able to compare exact values here
-                assert np.all(t[column] == d[column])
-                if t[column].dtype.isnative:
-                    assert d[column].dtype == t[column].dtype
-                else:
-                    assert d[column].dtype == t[column].dtype.newbyteorder()
-
-        # Regression test for astropy/astropy#1156 - the following code gave a
-        # ValueError: Big-endian buffer not supported on little-endian
-        # compiler. We now automatically swap the endian-ness to native order
-        # upon adding the arrays to the data frame.
-        # Explicitly testing little/big/native endian separately -
-        # regression for a case in astropy/astropy#11286 not caught by #3729.
-        d[["<i4", ">i4"]]
-        d[["<f4", ">f4"]]
-
-        t2 = table.Table.from_df(d)
-
-        for column in t.columns:
-            if column in ("u", "s"):
-                assert np.all(t[column] == t2[column])
-            else:
-                assert_allclose(t[column], t2[column])
-            if t[column].dtype.isnative:
-                assert t[column].dtype == t2[column].dtype
[…]
