+++ astropy/table/tests/test_df.py
@@ -157,9 +157,15 @@ class TestGenericDataFrames:
     @pytest.mark.parametrize(
         "backend",
         [
-            pytest.param("pandas", marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")),
-            pytest.param("polars", marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars")),
-        ]
+            pytest.param(
+                "pandas",
+                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
+            ),
+            pytest.param(
+                "polars",
+                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
+            ),
+        ],
     )
     def test_simple(self, backend):
         """Test basic endianness and data type handling."""
@@ -203,9 +209,15 @@ class TestGenericDataFrames:
     @pytest.mark.parametrize(
         "backend",
         [
-            pytest.param("pandas", marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")),
-            pytest.param("polars", marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars")),
-        ]
+            pytest.param(
+                "pandas",
+                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
+            ),
+            pytest.param(
+                "polars",
+                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
+            ),
+        ],
     )
     @pytest.mark.parametrize("unsigned", ["u", ""])
     @pytest.mark.parametrize("bits", [8, 16, 32, 64])
@@ -224,9 +236,15 @@ class TestGenericDataFrames:
     @pytest.mark.parametrize(
         "backend",
         [
-            pytest.param("pandas", marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")),
-            pytest.param("polars", marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars")),
-        ]
+            pytest.param(
+                "pandas",
+                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
+            ),
+            pytest.param(
+                "polars",
+                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
+            ),
+        ],
     )
     def test_2d_columns(self, backend):
         """Test handling of multidimensional columns."""
@@ -243,6 +261,7 @@ class TestGenericDataFrames:
         else:
             # Polars supports multidimensional columns as Array type
             import polars as pl
+
             tp = t.to_df(backend)
             assert isinstance(tp["b"].dtype, pl.Array)
 
@@ -254,9 +273,15 @@ class TestGenericDataFrames:
     @pytest.mark.parametrize(
         "backend",
         [
-            pytest.param("pandas", marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")),
-            pytest.param("polars", marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars")),
-        ]
+            pytest.param(
+                "pandas",
+                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
+            ),
+            pytest.param(
+                "polars",
+                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
+            ),
+        ],
     )
     def test_mixin_columns(self, backend):
         """Test handling of astropy mixin columns."""
@@ -303,9 +328,15 @@ class TestGenericDataFrames:
     @pytest.mark.parametrize(
         "backend",
         [
-            pytest.param("pandas", marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas")),
[…]
