### 2025-08-12 eae8136
Doc build bug fixes
