### 2025-08-12 dd8e08d
Doc build bug fixes
