+++ astropy/table/tests/test_df.py
@@ -8,7 +8,12 @@ from astropy import table
 from astropy import units as u
 from astropy.table import MaskedColumn, Table
 from astropy.time import Time, TimeDelta
-from astropy.utils.compat.optional_deps import HAS_NARWHALS, HAS_PANDAS, HAS_POLARS
+from astropy.utils.compat.optional_deps import (
+    HAS_NARWHALS,
+    HAS_PANDAS,
+    HAS_POLARS,
+    HAS_PYARROW,
+)
 
 from .conftest import MIXIN_COLS
 
@@ -151,22 +156,26 @@ class TestPandasLegacy:
 
 
 @pytest.mark.skipif(not HAS_NARWHALS, reason="requires narwhals")
+@pytest.mark.parametrize(
+    "backend",
+    [
+        pytest.param(
+            "pandas",
+            marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
+        ),
+        pytest.param(
+            "polars",
+            marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
+        ),
+        pytest.param(
+            "pyarrow",
+            marks=pytest.mark.skipif(not HAS_PYARROW, reason="requires pyarrow"),
+        ),
+    ],
+)
 class TestGenericDataFrames:
     """Test generic DataFrame functionality that should work across backends."""
 
-    @pytest.mark.parametrize(
-        "backend",
-        [
-            pytest.param(
-                "pandas",
-                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
-            ),
-            pytest.param(
-                "polars",
-                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
-            ),
-        ],
-    )
     def test_simple(self, backend):
         """Test basic endianness and data type handling."""
         t = table.Table()
@@ -206,19 +215,6 @@ class TestGenericDataFrames:
                 else:
                     assert t_dtype == t2_dtype
 
-    @pytest.mark.parametrize(
-        "backend",
-        [
-            pytest.param(
-                "pandas",
-                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
-            ),
-            pytest.param(
-                "polars",
-                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
-            ),
-        ],
-    )
     @pytest.mark.parametrize("unsigned", ["u", ""])
     @pytest.mark.parametrize("bits", [8, 16, 32, 64])
     def test_nullable_int(self, backend, unsigned, bits):
@@ -233,19 +229,6 @@ class TestGenericDataFrames:
         assert np.all(t2["col0"].mask == [False, True])
         assert np.all(t2["col0"] == c)
 
-    @pytest.mark.parametrize(
-        "backend",
-        [
-            pytest.param(
-                "pandas",
-                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
-            ),
-            pytest.param(
-                "polars",
-                marks=pytest.mark.skipif(not HAS_POLARS, reason="requires polars"),
-            ),
-        ],
-    )
     def test_2d_columns(self, backend):
         """Test handling of multidimensional columns."""
         t = table.Table()
@@ -270,19 +253,6 @@ class TestGenericDataFrames:
             tp = t.to_df(backend)
             assert isinstance(tp["b"].dtype, pl.Array)
 
-    @pytest.mark.parametrize(
-        "backend",
-        [
-            pytest.param(
-                "pandas",
-                marks=pytest.mark.skipif(not HAS_PANDAS, reason="requires pandas"),
-            ),
-            pytest.param(
-                "polars",
[…]
