### 2025-08-12 e02c987
Fix bug removing rows on an indexed and sliced table
