### 2025-08-12 06b2499
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
