### 2025-08-12 6864544
Squash me: Revert experimental changes
that did not work out
