+++ tox.ini
@@ -76,8 +76,7 @@ deps =
     fitsio: fitsio
 
     # Include scipy in double run to detect test pollution in scipy-only tests
-    double: numpy==2.2.2
-    double: scipy==1.15.2
+    double: scipy
 
     # The devdeps factor is intended to be used to install the latest developer version
     # or nightly wheel of key dependencies.
