### 2025-08-12 a499595
Squash me: Revert experimental changes
that did not work out
