### 2025-08-12 02170d4
Update IVOA UCD1+ controlled vocabulary from version 1.5 to 1.6 (#18483)
* Update IVOA UCD1+ controlled vocabulary from version 1.5 to 1.6

This update adds support for new atmospheric observation terms including
obs.atmos.wind, obs.atmos.humidity, obs.atmos.rain, obs.atmos.turbulence,
obs.atmos.turbulence.isoplanatic, and obs.atmos.water which are now
recognized when parsing UCDs with check_controlled_vocabulary=True.

Fixes #18452

* Fix formatting issues in UCD test

- Remove trailing whitespace and blank line whitespace
- Fix PLR0133 by using variable instead of inline string comparison
- 
[…]
