+++ astropy/io/votable/tests/test_ucd.py
@@ -69,3 +69,48 @@ def test_invalid_namespace():
 def test_invalid_word():
     with pytest.raises(ValueError):
         ucd.parse_ucd("-pho")
+
+
+def test_atmospheric_wind_ucd():
+    """Test parsing of atmospheric observation UCDs, specifically obs.atmos.wind.
+
+    This test addresses issue #18452 where obs.atmos.wind and other new
+    atmospheric terms from UCD1+ version 1.6 should be recognized.
+    """
+    # Test the specific case from the issue report
+    ucd_string = "phys.veloc;obs.atmos.wind;stat.mean"
+    result = ucd.parse_ucd(ucd_string, check_controlled_vocabulary=True)
+    expected = [
+        ("ivoa", "phys.veloc"),
+        ("ivoa", "obs.atmos.wind"),
+        ("ivoa", "stat.mean"),
+    ]
+    assert result == expected
+
+    # Test primary/quantified atmospheric terms (can be used standalone)
+    primary_atmos_terms = [
+        "obs.atmos.humidity",
+        "obs.atmos.rain",
+        "obs.atmos.turbulence.isoplanatic",
+        "obs.atmos.water",
+        "phys.temperature.dew",
+    ]
+
+    for term in primary_atmos_terms:
+        result = ucd.parse_ucd(term, check_controlled_vocabulary=True)
+        assert result == [("ivoa", term)]
+        assert ucd.check_ucd(term, check_controlled_vocabulary=True)
+
+    # Test secondary atmospheric terms (must be used with primary terms)
+    secondary_atmos_terms = [
+        "obs.atmos.wind",
+        "obs.atmos.turbulence",
+    ]
+
+    for term in secondary_atmos_terms:
+        # Test as secondary word (valid)
+        combined_ucd = f"phys.temperature;{term}"
+        result = ucd.parse_ucd(combined_ucd, check_controlled_vocabulary=True)
+        expected = [("ivoa", "phys.temperature"), ("ivoa", term)]
+        assert result == expected
+        assert ucd.check_ucd(combined_ucd, check_controlled_vocabulary=True)
