### 2025-08-12 457b038
Remove return statements from __init_subclass__()
The only value a `__init_subclass__()` method should return is `None`
and that can be done implicitly.
