### 2025-08-13 5f1c193
Updated use_nullable_int description to match tests
