### 2025-08-13 f7515d2
All tests passing, but no 1D support for PyArrow yet
