+++ astropy/table/tests/test_df.py
@@ -229,29 +229,36 @@ class TestGenericDataFrames:
         assert np.all(t2["col0"].mask == [False, True])
         assert np.all(t2["col0"] == c)
 
-    def test_2d_columns(self, backend):
-        """Test handling of multidimensional columns."""
+    @pytest.mark.parametrize("ndim", [1, 2, 3])
+    def test_nd_columns(self, backend, ndim):
+        # Add one since we want the dimension of each entry to be ndim
+        shape = (3,) * ndim
+        colshape = (10,) + shape
         t = table.Table()
-        t["a"] = [1, 2, 3]
-        t["b"] = np.ones((3, 2))
+        t["a"] = np.ones(colshape)
 
         if backend == "pandas":
-            # Pandas doesn't support multidimensional columns
             with pytest.raises(
-                ValueError, match="Cannot convert a table with multidimensional columns"
+                ValueError,
+                match="Cannot convert a table with multidimensional columns",
             ):
                 t.to_df(backend)
-        else:
-            # Polars supports multidimensional columns as Array type
-            import polars as pl
 
+        elif backend == "polars":
+            # Convert and check shape
             tp = t.to_df(backend)
-            assert isinstance(tp["b"].dtype, pl.Array)
+            assert tp["a"].dtype.shape == shape
 
-            # Test 3D as well
-            t["b"] = np.ones((3, 2, 2))
-            tp = t.to_df(backend)
-            assert isinstance(tp["b"].dtype, pl.Array)
+            # Round-trip conversion
+            t2 = table.Table.from_df(tp)
+            assert t2["a"].shape == colshape
+
+        elif backend == "pyarrow":
+            with pytest.raises(
+                ValueError,
+                match="only handle 1-dimensional arrays",
+            ):
+                t.to_df(backend)
 
     def test_mixin_columns(self, backend):
         """Test handling of astropy mixin columns."""
@@ -305,13 +312,25 @@ class TestGenericDataFrames:
 
         tp = t.to_df(backend)
 
-        if backend == "pandas":
-            assert np.all(tp["tm"].isnull() == [False, True, False])
-            assert np.all(tp["dt"].isnull() == [False, True, False])
-        else:  # polars
-            assert np.all(tp["tm"].is_null().to_list() == [False, True, False])
-            assert np.all(tp["dt"].is_null().to_list() == [False, True, False])
+        match backend:
+            case "pandas":
+                tm_nulls = tp["tm"].isnull().to_list()
+                dt_nulls = tp["dt"].isnull().to_list()
+            case "polars":
+                tm_nulls = tp["tm"].is_null().to_list()
+                dt_nulls = tp["dt"].is_null().to_list()
+            case "pyarrow":
+                tm_nulls = tp["tm"].is_null().to_pylist()
+                dt_nulls = tp["dt"].is_null().to_pylist()
+            case _:
+                raise ValueError(f"Unknown backend: {backend}")
 
+        # Common assertion for all backends
+        expected_nulls = [False, True, False]
+        assert tm_nulls == expected_nulls
+        assert dt_nulls == expected_nulls
+
+        # Round-trip conversion
         t2 = table.Table.from_df(tp)
 
         assert np.all(t2["tm"].mask == tm.mask)
@@ -368,105 +387,113 @@ class TestGenericDataFrames:
                 else:
                     assert column.dtype.newbyteorder() == t2[name].dtype
 
+    def test_units(self, backend):
+        """Test handling of units in from_df conversion."""
[…]
