### 2025-08-13 4c8afff
Raise correct Error for PyArrow multi-d arrays (for now)
