+++ astropy/table/dataframes.py
@@ -86,10 +86,10 @@ class DataFrameConverter:
         """
         # Check for multidimensional columns
         badcols = [name for name, col in table.columns.items() if len(col.shape) > 1]
-        if badcols and backend_impl.is_pandas_like():
+        if badcols and (backend_impl.is_pandas_like() or backend_impl.is_pyarrow()):
             raise ValueError(
                 f"Cannot convert a table with multidimensional columns to a "
-                f"pandas-like DataFrame. Offending columns are: {badcols}\n"
+                f"pandas-like or pyarrow DataFrame. Offending columns are: {badcols}\n"
                 f"One can filter out such columns using:\n"
                 f"names = [name for name in tbl.colnames if len(tbl[name].shape) <= 1]\n"
                 f"tbl[names].to_pandas(...)"
