### 2025-08-13 d47aa8c
Remove unnecessary hasattr check for col.info.indices
