### 2025-08-13 ee15355
Add docs: AI-generated then updated by hand.
