+++ astropy/table/index.py
@@ -1,34 +1,188 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-"""Indexing for Table columns.
+"""
+Table Indexing Implementation for Astropy Tables
+================================================
 
-The Index class can use several implementations as its
-engine. Any implementation should implement the following:
+This module implements indexing functionality for astropy Table objects, enabling fast
+lookups and range queries on table columns. The indexing system uses various data
+structures to maintain sorted representations of table data for efficient searching.
 
-__init__(data, row_index) : initialize index based on key/row list pairs
-add(key, row) -> None : add (key, row) to existing data
-remove(key, data=None) -> boolean : remove data from self[key], or all of
-                                    self[key] if data is None
-shift_left(row) -> None : decrement row numbers after row
-shift_right(row) -> None : increase row numbers >= row
-find(key) -> list : list of rows corresponding to key
-range(lower, upper, bounds) -> list : rows in self[k] where k is between
-                               lower and upper (<= or < based on bounds)
-sort() -> None : make row order align with key order
-sorted_data() -> list of rows in sorted order (by key)
-replace_rows(row_map) -> None : replace row numbers based on slice
-items() -> list of tuples of the form (key, data)
+Overview
+--------
+The astropy Table class supports creating indexes on one or more columns to enable fast
+data retrieval operations. When an index is created on a column (or combination of
+columns), a sorted data structure is built that maps column values to their
+corresponding row numbers in the original table.
 
-Notes
------
-    When a Table is initialized from another Table, indices are
-    (deep) copied and their columns are set to the columns of the new Table.
+The indices actually exist only on the columns. Calling ``Table.add_index()`` appends a
+new `SlicedIndex` instance to the `col.indices` list for each of the indexed columns.
+Adding an index always adds a `SlicedIndex` to the indices, with the slice initially
+having ``start=0``, ``stop=len(col)-1`` and ``step=1``. The class has an ``original``
+attribute which is initially ``True`` indicating that the index is not sliced.
 
-    Column creation:
-    Column(c) -> deep copy of indices
-    c[[1, 2]] -> deep copy and reordering of indices
-    c[1:2] -> reference
-    array.view(Column) -> no indices
+The ``Table.indices`` property is a convenience method that returns a dynamically
+created `TableIndices` object that collects the indices from table columns.
+
+The ``Table.loc`` and ``Table.iloc`` properties return a new instance of
+`TableLoc(table)` or `TableILoc(table)` each time.
+
+Key Components
+--------------
+
+Index Classes:
+    - `Index`: Wraps different engines and provides core index functions.
+    - `SlicedIndex`: Contains an Index and handles original or sliced tables.
+
+Index Engines:
+    - `SortedArray`: Array-based sorted container implementation (default)
+    - `SCEngine`: SortedContainers-based implementation (when available)
+    - `BST`: Binary search tree implementation for sorted data (not for production)
+
+Query Interface Classes:
+    - `TableLoc`: Value-based row retrieval using ``.loc[]``.
+    - `TableLocIndices`: Return row indices (not table slices) for ``.loc_indices[]``
+    - `TableILoc`: Return rows in based on sorted index order for ``.iloc[]``
+
+Core Functionality
+------------------
+As a reminder of basic functionality.
+
+Creating Indexes:
+    Tables support adding indexes via `Table.add_index()`:
+
+    >>> t = Table({'a': [1, 3, 2, 5], 'b': ['x', 'y', 'z', 'w']})
+    >>> t.add_index('a')  # Single column index
+    >>> t.add_index(['a', 'b'])  # Multi-column composite index
+
+Value-Based Lookups with TableLoc:
[…]
