### 2025-08-13 00f472f
Factor out operators as pytest parameters in new tests, as suggested by @neutrinoceros, to avoid putting assertions in for loops (which makes it hard to see which op is failing).
