+++ astropy/nddata/mixins/tests/test_ndarithmetic.py
@@ -1,6 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 import itertools
+import operator
 from copy import deepcopy
 
 import numpy as np
@@ -29,6 +30,15 @@ class StdDevUncertaintyUncorrelated(StdDevUncertainty):
     @property
     def supports_correlated(self):
         return False
+
+
+# Correspondence between NDArithmetic & Python method/function names:
+operator_mapping = {
+    "add" : operator.add,
+    "subtract" : operator.sub,
+    "multiply" : operator.mul,
+    "divide" : operator.truediv,
+}
 
 
 # Test with Data covers:
@@ -1393,28 +1403,16 @@ def test_nddata_bitmask_arithmetic():
         np.array(2.0, dtype=np.float64),
     ],
 )
-def test_arithmetics_dtypes_with_scalar(data1, data2):
-    # Addition
-    out1 = data1.add(data2)
-    ref1 = data1.data + data2
+@pytest.mark.parametrize(("meth", "op"), operator_mapping.items())
+def test_arithmetics_dtypes_with_scalar(data1, data2, meth, op):
 
-    # Subtraction
-    out2 = data1.subtract(data2)
-    ref2 = data1.data - data2
-
-    # Multiplication
-    out3 = data1.multiply(data2)
-    ref3 = data1.data * data2
-
-    # Division
-    out4 = data1.divide(data2)
-    ref4 = data1.data / data2
+    out = getattr(data1, meth)(data2)
+    ref = op(data1.data, data2)
 
     # Enforce the same behaviour as NumPy, rather than fixed behaviour:
-    for ndd, ref in zip((out1, out2, out3, out4), (ref1, ref2, ref3, ref4)):
-        assert ndd.data.shape == ref.shape
-        assert ndd.data.dtype == ref.dtype
-        assert_array_equal(ndd.data, ref)
+    assert out.data.shape == ref.shape
+    assert out.data.dtype == ref.dtype
+    assert_array_equal(out.data, ref)
 
 
 # Covers adding scalar quantity matching non-default dtypes:
@@ -1435,20 +1433,23 @@ def test_arithmetics_dtypes_with_scalar(data1, data2):
         ),
     ],
 )
-def test_add_quantity_matching_dtype(data1, data2):
-    # Addition
-    out1 = data1.add(data2)
-    assert_array_equal(out1.data, data1.data + data2.value)
+@pytest.mark.parametrize(
+    ("meth", "op"),
+    (
+        (k, v) for k, v in operator_mapping.items() if k in ("add", "subtract")
+    ),
+)
+def test_add_quantity_matching_dtype(data1, data2, meth, op):
 
-    # Subtraction
-    out2 = data1.subtract(data2)
-    assert_array_equal(out2.data, data1.data - data2.value)
+    out = getattr(data1, meth)(data2)
+    ref = op(data1.data, data2.value)
 
-    for ndd in [out1, out2]:
-        assert ndd.data.shape == data1.data.shape
-        assert ndd.data.dtype == data1.data.dtype
+    assert out.data.shape == data1.data.shape
+    assert out.data.dtype == data1.data.dtype  # expect no change in this case
+    assert_array_equal(out.data, ref)
 
 
+# Covers scaling with units and non-default dtypes:
 @pytest.mark.parametrize(
     "data1",
     [
@@ -1467,19 +1468,20 @@ def test_add_quantity_matching_dtype(data1, data2):
         np.array(2.0),
     ],
 )
-def test_scale_dtypes_with_units(data1, data2):
-    # Multiplication
-    out1 = data1.multiply(data2)
-    ref1 = data1.data * data2
[…]
