### 2025-08-13 f59ee22
Use actual NumPy scalars, rather than just 0-d arrays, which aren't quite the same thing.
