### 2025-08-13 fb24917
Eliminate assertions in loops from pre-existing tests, as suggested by @neutrinoceros, by factoring out the different operators as pytest parameters.
