+++ astropy/nddata/mixins/tests/test_ndarithmetic.py
@@ -57,34 +57,25 @@ operator_mapping = {
         (np.arange(1000).reshape(20, 5, 10), np.ones((20, 5, 10)) * 3),
     ],
 )
-def test_arithmetics_data(data1, data2):
+@pytest.mark.parametrize(("meth", "op"), operator_mapping.items())
+def test_arithmetics_data(data1, data2, meth, op):
     nd1 = NDDataArithmetic(data1)
     nd2 = NDDataArithmetic(data2)
 
-    # Addition
-    nd3 = nd1.add(nd2)
-    assert_array_equal(data1 + data2, nd3.data)
-    # Subtraction
-    nd4 = nd1.subtract(nd2)
-    assert_array_equal(data1 - data2, nd4.data)
-    # Multiplication
-    nd5 = nd1.multiply(nd2)
-    assert_array_equal(data1 * data2, nd5.data)
-    # Division
-    nd6 = nd1.divide(nd2)
-    assert_array_equal(data1 / data2, nd6.data)
-    for nd in [nd3, nd4, nd5, nd6]:
-        # Check that broadcasting worked as expected
-        if data1.ndim > data2.ndim:
-            assert data1.shape == nd.data.shape
-        else:
-            assert data2.shape == nd.data.shape
-        # Check all other attributes are not set
-        assert nd.unit is None
-        assert nd.uncertainty is None
-        assert nd.mask is None
-        assert len(nd.meta) == 0
-        assert nd.wcs is None
+    nd = getattr(nd1, meth)(nd2)
+    assert_array_equal(op(data1, data2), nd.data)
+
+    # Check that broadcasting worked as expected
+    if data1.ndim > data2.ndim:
+        assert data1.shape == nd.data.shape
+    else:
+        assert data2.shape == nd.data.shape
+    # Check all other attributes are not set
+    assert nd.unit is None
+    assert nd.uncertainty is None
+    assert nd.mask is None
+    assert len(nd.meta) == 0
+    assert nd.wcs is None
 
 
 # Invalid arithmetic operations for data covering:
@@ -117,45 +108,27 @@ def test_arithmetics_data_invalid():
         (np.array(5), np.array(10) * u.s / u.h),
     ],
 )
-def test_arithmetics_data_unit_identical(data1, data2):
+@pytest.mark.parametrize(("meth", "op"), operator_mapping.items())
+def test_arithmetics_data_unit_identical(data1, data2, meth, op):
     nd1 = NDDataArithmetic(data1)
     nd2 = NDDataArithmetic(data2)
 
-    # Addition
-    nd3 = nd1.add(nd2)
-    ref = data1 + data2
+    nd = getattr(nd1, meth)(nd2)
+    ref = op(data1, data2)
     ref_unit, ref_data = ref.unit, ref.value
-    assert_array_equal(ref_data, nd3.data)
-    assert nd3.unit == ref_unit
-    # Subtraction
-    nd4 = nd1.subtract(nd2)
-    ref = data1 - data2
-    ref_unit, ref_data = ref.unit, ref.value
-    assert_array_equal(ref_data, nd4.data)
-    assert nd4.unit == ref_unit
-    # Multiplication
-    nd5 = nd1.multiply(nd2)
-    ref = data1 * data2
-    ref_unit, ref_data = ref.unit, ref.value
-    assert_array_equal(ref_data, nd5.data)
-    assert nd5.unit == ref_unit
-    # Division
-    nd6 = nd1.divide(nd2)
-    ref = data1 / data2
-    ref_unit, ref_data = ref.unit, ref.value
-    assert_array_equal(ref_data, nd6.data)
-    assert nd6.unit == ref_unit
[…]
