### 2025-08-14 2d01f79
Accept Ruff formatting suggestions.
