### 2025-08-14 ee182bd
Add option to further sanitize slices for SlicedLowLevelWCS based on data shape.
