+++ astropy/wcs/wcsapi/wrappers/tests/test_sliced_wcs.py
@@ -65,10 +65,14 @@ def test_invalid_slices():
         SlicedLowLevelWCS(WCS_SPECTRAL_CUBE, [None, None, 1000.100])
 
     with pytest.raises(IndexError):
-        SlicedLowLevelWCS(WCS_SPECTRAL_CUBE, [slice(-1, -2)], presliced_shape=(50, 60, 70))
+        SlicedLowLevelWCS(
+            WCS_SPECTRAL_CUBE, [slice(-1, -2)], presliced_shape=(50, 60, 70)
+        )
 
     with pytest.raises(IndexError):
-        SlicedLowLevelWCS(WCS_SPECTRAL_CUBE, [slice(55, 50)], presliced_shape=(50, 60, 70))
+        SlicedLowLevelWCS(
+            WCS_SPECTRAL_CUBE, [slice(55, 50)], presliced_shape=(50, 60, 70)
+        )
 
 
 @pytest.mark.parametrize(
