### 2025-08-14 c9c2d13
Added extra sanitization case to sanitize_slices.
Catches cases when negative slice start or index values are beyond extent of array associated
with SlicedLowLevelWCS. Also added more test coverage.
