### 2025-08-14 d3523bf
Remove two `ErfaAstromInterpolator` static methods
The removed static methods were only called by the `apco()` method and
their functionality is simpler to achieve using a local function defined
with `functools.partial()`.
