### 2025-08-15 d132071
CLN: simplify argument parser settings
