### 2025-08-15 958a609
CLN: simplify argument parser settings
