### 2025-08-15 4b586be
Add additional cautious remarks on timing accuracy in multiple places
