### 2025-08-15 e7eefee
Move duplicated code to a new function
