### 2025-08-15 f5eb88f
Return a `NamedTuple` from a helper function
Returning a `NamedTuple` with named attributed allows eliminating local
variables that store the function's output because the variable names
are no longer needed for documentating the code.
