### 2025-08-16 3164fb7
Activate as many doctests and feasible in narrative unified I/O interface
doctests with dependencies (hdf5 and parquet) are set up, but still skipped due to
https://github.com/scientific-python/pytest-doctestplus/issues/291
