### 2025-08-16 bd2bdce
Activate doctests for hdf5 and parquet with individual `doctest-requires` guards
closes #13265
