+++ docs/io/unified_table_hdf5.rst
@@ -1,5 +1,3 @@
-.. doctest-skip-all
-
 .. _table_io_hdf5:
 
 HDF5
@@ -28,21 +26,31 @@ you can do:
 .. testsetup::
     >>> from astropy.table import QTable
     >>> tab = QTable({'col1': [1, 2, 3], 'col2': ['a', 'b', 'c']})
-    >>> tab.write('observations.hdf5', path='data')
-    >>> tab.write('observations.hdf5', path='group/data', append=True)
+    >>> from astropy.utils.compat.optional_deps import HAS_H5PY
+    >>> if HAS_H5PY:
+    ...     tab.write('observations.hdf5', path='data')
+    ...     tab.write('observations.hdf5', path='group/data', append=True)
 
->>> from astropy.table import QTable
->>> t = QTable.read('observations.hdf5', path='data')
+.. doctest-requires:: h5py
 
-To read a table nested in a group in the HDF5 file, you can do::
+    >>> from astropy.table import QTable
+    >>> t = QTable.read('observations.hdf5', path='data')
+
+To read a table nested in a group in the HDF5 file, you can do:
+
+.. doctest-requires:: h5py
 
     >>> t = QTable.read('observations.hdf5', path='group/data')
 
-To write a table to a new file, the path should also be specified::
+To write a table to a new file, the path should also be specified:
+
+.. doctest-requires:: h5py
 
     >>> t.write('new_file.hdf5', path='updated_data')
 
-It is also possible to write a table to an existing file using ``append=True``::
+It is also possible to write a table to an existing file using ``append=True``:
+
+.. doctest-requires:: h5py
 
     >>> t.write('observations.hdf5', path='updated_data', append=True)
 
@@ -52,15 +60,11 @@ file that has multiple datasets, use *both* the ``overwrite=True`` and
 ``append=True`` arguments.
 
 Finally, when writing to HDF5 files, the ``compression=`` argument can be
-used to ensure that the data is compressed on disk::
+used to ensure that the data is compressed on disk:
+
+.. doctest-requires:: h5py
 
     >>> t.write('new_file.hdf5', path='updated_data', compression=True, overwrite=True)
-
-.. testcleanup::
-    >>> import os
-    >>> os.remove('new_file.hdf5')
-    >>> os.remove('observations.hdf5')
-
 
 ..
   EXAMPLE END
@@ -87,9 +91,11 @@ To enable storing all table and column metadata to the HDF5 file, call
 the ``write()`` method with ``serialize_meta=True``. This will store metadata
 in a separate HDF5 dataset, contained in the same file, which is named
 ``<path>.__table_column_meta__``. Here ``path`` is the argument provided in
-the call to ``write()``::
+the call to ``write()``:
 
-    >>> t.write('observations.hdf5', path='data', serialize_meta=True)
+.. doctest-requires:: h5py
+
+    >>> t.write('observations.hdf5', path='data', serialize_meta=True, overwrite=True)
 
 The table metadata are stored as a dataset of strings by serializing the
 metadata in YAML following the `ECSV header format
@@ -105,3 +111,9 @@ to HDF5 tables that contain :ref:`mixin_columns` such as `~astropy.time.Time` or
 .. note::
     Certain kind of metadata (e.g., numpy object arrays) cannot be serialized correctly
     using YAML.
+
+.. testcleanup::
+    >>> import os
+    >>> if HAS_H5PY:
+    ...     os.remove('new_file.hdf5')
+    ...     os.remove('observations.hdf5')
