### 2025-08-18 c2858da
Revert "Bump PyArrow"
This reverts commit b2e939fc5a9a7c05176800fd95bd47a359fac4b2.
