### 2025-08-18 33c0007
fixed separability_matrix for CompoundModel with fix_inputs (issue #18523)
