### 2025-08-18 fc68633
Use more files included in astropy for examples instead of secretly generated files
