### 2025-08-18 2244266
fixed changelog grammar
