### 2025-08-18 143a737
Add a new Time ``GalexSec`` for the GALEX satelite (#18330)
* Add a new Time ``GalexSec`` for the GALEX satelite

We've had Chandta in astropy for a long time, but I think it's time to add more observatories that use similar schemes - but with different scales and zeropoints.

* Address code review

- Update code comment about time accuracy
- Improvements for formatting and tests

Co-authored-by: Tom Aldcroft <taldcroft@gmail.com>

* Add additional cautious remarks on timing accuracy in multiple places

---------

Co-authored-by: Tom Aldcroft <taldcroft@gmail.com>
