### 2025-08-18 9c3517d
Fix bug where coord frame fails serializing to YAML
