+++ astropy/io/misc/yaml.py
@@ -55,6 +55,7 @@ Examples
 """
 
 import base64
+import itertools
 
 import numpy as np
 import yaml
@@ -328,15 +329,23 @@ for cls, tag in (
     AstropyDumper.add_multi_representer(cls, _quantity_representer(tag))
     AstropyLoader.add_constructor(tag, _quantity_constructor(cls))
 
-for cls in list(coords.representation.REPRESENTATION_CLASSES.values()) + list(
-    coords.representation.DIFFERENTIAL_CLASSES.values()
-):
+# Add representations, differentials, frames defined in astropy.
+cls_coords = []
+cls_coords.extend(
+    cls
+    for cls in itertools.chain(
+        coords.representation.REPRESENTATION_CLASSES.values(),
+        coords.representation.DIFFERENTIAL_CLASSES.values(),
+    )
+    if cls.__name__ in coords.representation.__all__
+)
+cls_coords.extend(coords.frame_transform_graph.frame_set)
+
+for cls in cls_coords:
     name = cls.__name__
-    # Add representations/differentials defined in astropy.
-    if name in coords.representation.__all__:
-        tag = "!astropy.coordinates." + name
-        AstropyDumper.add_multi_representer(cls, _quantity_representer(tag))
-        AstropyLoader.add_constructor(tag, _quantity_constructor(cls))
+    tag = "!astropy.coordinates." + name
+    AstropyDumper.add_multi_representer(cls, _quantity_representer(tag))
+    AstropyLoader.add_constructor(tag, _quantity_constructor(cls))
 
 
 def load(stream):
