### 2025-08-18 833e010
move doctest requirements to pytest.toml
