### 2025-08-18 afff97c
move doctest requirements to pytest.toml
