+++ docs/io/unified_table_parquet.rst
@@ -31,15 +31,11 @@ Examples
     >>> if HAS_PANDAS and HAS_PYARROW:
     ...     t.write('observations.parquet')
 
-To read a table from a Parquet file named ``observations.parquet``, you can do:
-
-.. doctest-requires:: pandas pyarrow
+To read a table from a Parquet file named ``observations.parquet``, you can do::
 
     >>> t = Table.read('observations.parquet')
 
-To write a table to a new file, simply do:
-
-.. doctest-requires:: pandas pyarrow
+To write a table to a new file, simply do::
 
     >>> t.write('new_file.parquet')
 
@@ -49,15 +45,11 @@ overwriting existing files.
 One big advantage of the Parquet files is that each column is stored independently,
 and thus reading a subset of columns is fast and efficient.  To find out which
 columns are stored in a table without reading the data, use the ``schema_only=True``
-as shown below. This returns a zero-length table with the appropriate columns:
-
-.. doctest-requires:: pandas pyarrow
+as shown below. This returns a zero-length table with the appropriate columns::
 
     >>> schema = Table.read('observations.parquet', schema_only=True)
 
-To read only a subset of the columns, use the ``include_names`` and/or ``exclude_names`` keywords:
-
-.. doctest-requires:: pandas pyarrow
+To read only a subset of the columns, use the ``include_names`` and/or ``exclude_names`` keywords::
 
     >>> t_sub = Table.read('observations.parquet', include_names=['mjd', 'airmass'])
 
