### 2025-08-22 #251 @eteq
@lwasser - I think we want to move ahead with this review but just wanted to confirm a few things:

This package is of course a *much* larger package than most pyOpenSci reviews. We (the Astropy Coordination Committee) are guessing that the expected level of review is comparable to what sunpy got in #147 - i.e., a bit more high-level of a review and not e.g. expecting the reviewer to confirm *every single public function* has complete docstrings.  Assuming that's what you're thinking as well we are happy to have the process continue.  But we do not want to create undue burden on some poor reviewer or editor (nor on our own maintainers given they aren't necessarily all signed up for this).  So please advise,  if that sounds right to you we are happy to move forward with this.
