### 2025-09-03 #251 @lwasser
@eteq, my apologies for the delay in responding here. I took a vacation!

In terms of moving forward, yes, we would definitely expect that the Astropy review would be more high-level and similar to what SunPy went through. I think the next step will be to figure out who can lead this as editor. Will @dhomeier and @hamogu have a conflict of interest (I am not familiar with who works on what parts of Astropy!). cc: @pllim

Let's figure that out together and then we can begin to move things forward!
