### 2025-09-04 #251 @hamogu
I think this is best served by an editor other than me and @dhomeier . I'm not too worried about a conflict of interest, but the package is just so big and so established, that the most beneficial would be a really outside perspective, both for the editor and the reviewers. Every potential reviewer within the astropy community has worked on that package themselves in some form other the years (weather through code contributions or community discussions or some other means) and the packaging standards, layout, testing, docs etc., are what is to be considered the standard that other atrocities package should looks for.

So, if this review has any use, I think it's going to be that of an outside perspective. On the other hand, no specific astronomy skills are needed - we don't expect the reviewer to scrutinize e.g. how tile-compressed fits arrays with BZERO metadata are handled or how the redshift and distance of a galaxy depend on the exact value of the Hubble constant - those things are vetted carefully within the community already.
