### 2025-09-04 #251 @pllim
> what is to be considered the standard that other atrocities package should looks for

If no one else, I am sure @Cadair will not be shy to point those out... 😆
