### 2025-09-04 #251 @lwasser
Ok! I think finding the right editor and reviewers will be the tricky part here. I'll sync with @eliotwrobson on how to move this one forward!!

Because most of the astro community has likely contributed to this package, perhaps we look for one reviewer who is more focused on the package and usability, and another who, if they have contributed, has made small contributions (not a maintainer). Does that sound reasonable to you all?
