### 2025-10-09 #251 @lwasser
Hi everyone! 👋🏻 We are close to being able to kick off the review here for Astropy. First, we will run the EiC checks, so please stay tuned for more on that. We also already have an editor lined up, @crhea93, but we need to complete EiC checks first and we are at the beginning of our EiC rotation this month..

In the meantime, I just realized - WOW, Astropy has a HUGE maintainer team. Let's discuss how to list authors/maintainers meaningfully on the pyOpenSci website. Perhaps we link to the team landing page?

I do want to ensure that everyone gets credit for their work on this. We won't be able to fit everyone on our package listing page [here](https://www.pyopensci.org/python-packages.html):

But we could include a link if the list is a certain size in our code, and then you could add as many as you wanted to (or just use the link above).
