### 2025-10-09 #251 @pllim
I think https://www.astropy.org/team is fine. No need to duplicate effort. We update the team page regularly and contributor list is updated every release.
