### 2025-10-13 #251 @yeelauren
Hey All 👋
Just catching up on the conversation here, will have EiC checks completed this week!
