### 2025-10-24 #251 @yeelauren
- [x] **Installation** The package can be installed from a community repository such as PyPI (preferred), and/or a community channel on conda (e.g., conda-forge, bioconda).
- [x] **Documentation** The package has sufficient online documentation to allow us to evaluate the package's function and scope *without installing the package*. This includes:
  - [x] Short quickstart tutorials that help a user understand how to use the package and what it can do for them.
  - [x] API documentation (documentation for your code's functions, classes, methods, and attributes): this includes clearly written docstrings with variables defined using a standard docstring format.
  - [x] **README** The package has a `README.md` file with a clear explanation of what the package does, instructions on how to install it, and a link to development instructions.

**Optional:** Let projects know that it's a great idea for projects to have a .github repository for the project organization where they can host a commonly used LICENSE, Code of Conduct, and even a YAML file with label definitions. These items will then be automatically applied to every repository in the organization to ensure consistency (but can be customized within repos too). The [SunPy project](https://github.com/sunpy/.github/) has a great example of this.

---
Thank you, authors, in advance for setting aside five to ten minutes to do this. It truly helps our organization. :raised_hands:
---

👋  Astropy is in amazing shape to move on in the review process. Some minor suggestions:

- Contributing guide on github is more complete here https://github.com/astropy/astropy?tab=contributing-ov-file  than on the website https://docs.astropy.org/en/stable/index_dev.html
- could you link to https://docs.astropy.org/en/latest/index_dev.html directly ?
- suggest adding code of conduct as a .md file  to github from here https://www.astropy.org/code_of_conduct.html
