### 2025-10-24 #251 @pllim
Hello and thank you! Some follow-up below:

> Contributing guide

https://docs.astropy.org/en/stable/index_dev.html is purposefully brief. You may notice that at the bottom, it points to https://docs.astropy.org/en/latest/index_dev.html and if you follow that link, you will notice that it is way more complete than the stable version. This is because we ran into problems of users accidentally following outdated practices because stable version only gets updated every release, while the development stuff may change in between.

> could you link to ... directly?

Are you suggesting that I clear all the contents of https://github.com/astropy/astropy/blob/main/CONTRIBUTING.md and just throw in a link to https://docs.astropy.org/en/latest/index_dev.html ? If not, please clarify.

> suggest adding code of conduct as a .md file to github from here ...

I am afraid I do not understand this request. We already have https://github.com/astropy/astropy/blob/main/CODE_OF_CONDUCT.md and it does link to https://www.astropy.org/code_of_conduct.html . The latter is the canonical source, though at the moment, we are also in the midst of switching over to using NumFOCUS version (https://github.com/astropy/astropy.github.com/pull/669).
