### 2025-10-30 #251 @crhea93
## Editor response to review:

---

## Editor comments
Greetings @pllim! Congratulations on this monumental (and influential) package. As editor, it will be my job to find reviewers and guide them through the process. Additionally, I am here to help YOU with anything that might come up during this process. I look forward to the review!

~~I've started to reach out to potential reviewers, so hopefully the review can begin in earnest in the upcoming weeks.~~

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: December 15th, 2025
---

Reviewers: @adityapt, @egbdfx
Due date: December 15th, 2025
