### 2025-10-30 #251 @pllim
Code of conduct -- It is shared across the project, so we want to avoid the risk of every repo having their own complete copies and then go out of sync. Hopefully the way we do it right now is not a deal breaker for review.

Contributing guide -- Do you still want me to do something with https://github.com/astropy/astropy/blob/main/CONTRIBUTING.md ?
