### 2025-10-31 #251 @yeelauren
@pllim That makes sense to me - just to be clear there's no action required nor a deal breaker :)
Looking forward to the next step in the process soon 🥳
