### 2025-11-13 #251 @adityapt
Thanks @crhea93 ! I am going to create my review checklist here and will update it as I keep reveiwing the package.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

- [x] **Functionality:** Any functional claims of the software been confirmed. (core functionality - time, coordinates, units, modeling, FITS I/O - all these work)
- [x] **Automated tests:** On Python 3.11, Astropy 7.2.0 + NumPy + Matplotlib 3.7.2

- [x] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**  (I didn't reach out to the author for any bugs as there were not any).

---

The package seems to work as documented and expected. I was able to run the tests in my local, there were 2 minimal errors due to too many inputs being filtered. There are no performance and functionality claims that are quantified for me to validate, so I am not checking those boxes. Since 99% of the tests passed, I am checking it as all tests passed. I don't see a paper.md for JOSS submission, so I am not checking those. Overall, it's a great package that is well maintained by the community.

============================================ short test summary info ============================================
FAILED astropy-env/lib/python3.11/site-packages/astropy/time/tests/test_precision.py::test_datetime_timedelta_sum[tcg] - hypothesis.errors.FailedHealthCheck: It looks like this test is filtering out a lot of inputs. 9 inputs were...
FAILED astropy-env/lib/python3.11/site-packages/astropy/time/tests/test_precision.py::test_datetime_timedelta_sum[tdb] - hypothesis.errors.FailedHealthCheck: It looks like this test is filtering out a lot of inputs. 9 inputs were...
============== 2 failed, 25079 passed, 3758 skipped, 217 xfailed, 16 warnings in 215.07s (0:03:35) ==============
