### 2025-12-06 astropy/astropy#19032 @adityapt
Using 7.2.0 -
### Description

Hi All, I'm reviewing the astropy package 7.2.0 version - [review](https://github.com/pyOpenSci/software-submission/issues/251). There seems to be some inconsistency in the documentation - https://docs.astropy.org/en/stable/constants/index.html:

```python
from astropy import physical_constants, astronomical_constants
physical_constants.set('codata2010')
```

This code gives me a runtime eror:
```
RuntimeError: astropy.units is already imported
RuntimeError                              Traceback (most recent call last)
File <command-8207961779670862>, line 2
      1 from astropy import physical_constants, astronomical_constants
----> 2 physical_constants.set('codata2010')

File /local_disk0/.ephemeral_nfs/envs/pythonEnv-c0887ed7-5768-4df1-8e4f-8d286b59d504/lib/python3.12/site-packages/astropy/__init__.py:136, in base_constants_version.set(cls, value)
    133 import sys
    135 if "astropy.units" in sys.modules:
--> 136     raise RuntimeError("astropy.units is already imported")
    137 if "astropy.constants" in sys.modules:
    138     raise RuntimeError("astropy.constants is already imported")

RuntimeError: astropy.units is already imported
```
I also get another error when I run this code:
```python
from astropy import physical_constants, astronomical_constants
astronomical_constants.set('iau2012')
astronomical_constants.get()
```
```
RuntimeError: astropy.units is already imported
RuntimeError                              Traceback (most recent call last)
File <command-8207961779670863>, line 4
      1 from astropy import physical_constants, astronomical_constants
      2 # physical_constants.set('codata2010')
      3 # physical_constants.get()
----> 4 astronomical_constants.set('iau2012')
      5 astronomical_constants.get()

File /local_disk0/.ephemeral_nfs/envs/pythonEnv-c0887ed7-5768-4df1-8e4f-8d286b59d504/lib/python3.12/site-packages/astropy/__init__.py:136, in base_constants_version.set(cls, value)
    133 import sys
    135 if "astropy.units" in sys.modules:
--> 136     raise RuntimeError("astropy.units is already imported")
    137 if "astropy.constants" in sys.modules:
    138     raise RuntimeError("astropy.constants is already imported")

RuntimeError: astropy.units is already imported
```
### Expected behavior

No runtime error

### How to Reproduce

1. Get package from 'pip install astropy' version is 7.2.0
2. Then run '...'
3. An error occurs.

```python
from astropy import physical_constants, astronomical_constants
astronomical_constants.set('iau2012')
astronomical_constants.get()
physical_constants.set('codata2010')
```

### Versions

```python
import astropy
astropy.system_info()
```
```
platform
--------
platform.platform() = 'Linux-5.15.0-1096-aws-x86_64-with-glibc2.39'
platform.version() = '#103-Ubuntu SMP Mon Oct 13 19:46:03 UTC 2025'
platform.python_version() = '3.12.3'

packages
--------
astropy              7.2.0
numpy                2.1.3
scipy                1.15.1
matplotlib           3.10.0
pandas               2.2.3
pyerfa               2.0.1.5
```
