### 2025-12-06 astropy/astropy#19032 @adityapt
Looks like they were automatically imported. Didn't read the full documentation. Can we use a logger instead of Runtimeerror in the future?
