### 2025-12-08 #251 @pllim
Hello and thank you!

Astropy already published papers in other venues (ref: https://github.com/astropy/astropy/blob/main/astropy/CITATION) and therefore I don't think we need a `paper.md`?
