### 2025-12-08 #251 @crhea93
Hi there!
@adityapt thank you very much for the review! As for the paper, that is only necessary if astropy was planning on a JOSS submission :)
