### 2025-12-08 #251 @crhea93
@pllim hmm I see in the original submission that you did in fact mark that you would like to apply to JOSS as well. Is this the case?
