### 2025-12-08 #251 @pllim
Ooops, I guess I did. Let me [ask the rest of Astropy](https://groups.google.com/g/astropy-dev/c/obwxJC_b4Fo) and get back to you. Thanks for your patience!
