### 2025-12-08 #251 @crhea93
Not a problem at all! If you do decide to go ahead with a JOSS submission, that will be completed after the pyopensci part :)
