### 2025-12-08 #251 @pllim
> respond to @adityapt's comments

I responded at astropy/astropy#19032 and will circle back on `paper.md` later. Did I miss anything else?
