### 2025-12-12 #251 @pllim
Hello again! Astropy has decided not to pursue JOSS listing, therefore I have unchecked the "Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)?" box in my original post above. Thank you for your patience.

Please let me know what is the next step.
