### 2025-12-13 #251 @egbdfX
@pllim @crhea93 Thank you for the submission. My comments are provided below.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

The package functions as described in the documentation and behaves as expected. I was able to run the test suite locally without encountering any erros (```26138 passed, 3861 skipped, 220 xfailed, 17 warnings in 352.79s (0:05:52)```). Regarding performance, it is not practical to exhaustively test all of the functionality provided by Astropy; however, I ran several examples from the [tutorial page](https://learn.astropy.org/), and they worked as intended, so I have checked that box. Overall, this is a well-designed package that appears to be actively maintained by the community.
