### 2025-12-17 #251 @crhea93
---

🎉 astropy has been approved by pyOpenSci! Thank you @pllim  for submitting astropy and many thanks to @adityapt and @egbdfX  for reviewing this package! 😸 Also a huge shout out to the entire astropy team and all those who have been working on/maintaining this unbelievably useful packge. I know that personally completing my Ph.D. would have been an absolute nightmare without astropy.

I'd like to invite you all to fill out the post review surveys at the bottom of this comment. Additionally, @pllim I invite you to write a blog post to further highlight you package that will appear on our page. If this is of interest to you, I'll gladly provide you with additional details.

@pllim Please complete the tasks below at your earliest convenience!

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [x] Change the status tag of the issue to `6/pyOS-approved6 🚀🚀🚀`.  (You can remove all other labels; If the the package is moving on to JOSS then add a JOSS label)
- [x] Update the header of this issue with the date accepted, the version accepted, and the accepted version DOI (from Zenodo) to the header of this issue.
- [x] Invite the package maintainer(s) and both reviewers to the pyOpenSci Slack.
- [ ]  If the review is complete (not submitting to JOSS), close the issue when all items on this checklist are complete. If it's moving to JOSS, please close the issue once the JOSS review is complete and you've performed the task below.

---
