### 2025-12-18 #251 @pllim
My checklist

- [x] Activate [Zenodo](https://zenodo.org/) watching the repo if you haven't already done so. (It had been activated years ago.)
- [x] Tag and create a release to create a Zenodo version and DOI. (See https://zenodo.org/records/17756022)
- [ ] Add the badge for pyOpenSci peer-review to the README.md (Being done via https://github.com/astropy/astropy/pull/19065)
- [x] Please fill out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7). All maintainers and reviewers should fill this out. (Done!)
