### 2025-12-18 #251 @pllim
> write a blog post

I don't think astropy write blogs but I can [ask](https://groups.google.com/g/astropy-dev/c/obwxJC_b4Fo)...
