### 2025-12-18 #251 @pllim
I feel like it is more proper to merge the badge PR https://github.com/astropy/astropy/pull/19065 after this is closed. What do you think? Thanks, all!
