### 2026-01-20 #251 @pllim
Huh, somehow astropy is already showing up at https://www.pyopensci.org/python-packages.html even though this issue is still open.
