### 2026-01-21 #251 @hamogu
@pllim The script for the website uses the labels, not the open/closed status of the issue. Issues with label "6 - approved" are listed. It's common that issues stay open longer, in particular if the package also wants a JOSS article. The JOSS review only starts after PyOS accepts a package, so we don't want to wait for the JOSS formalities to go through before listing on our webpage.
