### 2026-01-22 #251 @pllim
Do we have a follow-up issue to follow up on the wrong listed info as mentioned in https://github.com/pyOpenSci/software-submission/issues/251#issuecomment-3774181142 ? Thanks!
