### 2026-01-23 #251 @lwasser
Hi friends @pllim, Yesterday, after chatting with @crhea93, I realized I had broken the astropy page while debugging another page. The `Hello there` was me debugging some messy liquid. That should all be fixed now. My apologies for that miss when I merged! I was focused on another page not astropy!

The list of maintainers on that card comes from the YAML in this issue. Right now, this is what it looks like (please scroll up to the top) .

SO, I know you have a huge team for astropy, and my question is, should we just add the team link and the head maintainers? or what would be best here? This is the first package that has such a large maintainer team (which we love!!). But our current script parses that field for gh usernames and commas and so it doesn't know how to handle what is currently there!

Let's brainstorm what might work best. And we can do that here or in Slack - I'm sorry I've missed pings here.
