### 2026-01-23 #251 @lwasser
So if you were to update the maintainers today, and we can manually retrigger a build, it would fix the card. But if we want a link or something that is not a github username list, we will need to update our build to accept that
