### 2026-01-23 #251 @pllim
To be consistent with astropy citation, I would go with "Astropy Collaboration" but that is being discussed over at https://github.com/astropy/astropy.github.com/pull/700 (cc @hamogu and @dhomeier)

Regardless, it is best not to list people individually.
